==> 00-complete.Rmd <==
---
title: "PROJECT: ELISA standardization"
author: "Andree Valle Campos"
date: '`r Sys.Date()`'
output: 
  html_document:
#  pdf_document:
#  html_notebook:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
    code_folding: "hide"
    number_sections: TRUE
#    df_print: kable
#    fig_caption: true
#  documentclass: report
bibliography: analysis/SeroMarker.bib
csl: analysis/american-medical-association.csl
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,     # CHANGE TO FALSE IN PDF_DOCUMENTS
                      warning = FALSE) 
#knitr::opts_knit$set(root.dir = '../.')
options(width = 110) # expand limits of CONSOLE output  # CHANGE TO DEFAULT in PDF
# FOR PDF CHANGE: html_document, toc_float, number_sections, echo, width
# FOR PDF add new page USE: \newpage
```

```{r, results='hide', message=FALSE, eval=TRUE}
require(Hmisc) # to call markupSpecs
mu <- markupSpecs$html # expand limits of HTML SCREEN output to use an entire wide screen
# with the following hidden command (<code>r mu$widescreen()</code>).
#mu$widescreen()
```
`r mu$widescreen()`

```{r child = '00-appendix.Rmd'}
```

```{r child = '02-dtfilter.Rmd'}
```

```{r child = '03-covariat.Rmd'}
```

```{r child = '04-epidemio.Rmd'}
```

# References
