==> 03-covariat.Rmd <==
# COVARIATE UNIFICATION {#covar}

## Aim

* Principal: 
    - Generate a unified covariate database for the following statistical analysis.

* Secondary: 
    - Generate a Quality Control report comparing UNAP and NAMR measurements on microscopy and PCR.

## Results

- __non-stardardized DB names__ were cronologically standardized. [here](#bashcov)
- __prior and update__ information was summarized in a __diagram__. [here](#priorupgraph)
- `unap0` was created as a **merged** DB of covariates. [here](#merge-covariates)
- `unap0` was **merged** with both `std.nty` and `std.uty` DB's of **mean.OD** and **Ab.units** 
per specie previously obtained. [here](#merge-ab-units)
    + At the moment, **`std.uty.cov`** is the most convenient dataset.
- **Consensus variable** was generated for `micro` and `pcr` giving preference to **NAMR** measurement. [here](#generate-consensus)
- A **QC report** shows __97.5% similarity__ for `microscopy` and __73.4%__ for `pcr`. [here](#qc-report)
    + Percentages increases assuming independent observations with regard to the `specie`. [here](#tidyqcrev)

## To Do

- In order to keep using `std.nty` as a tidy dataset 
    + A [relational data](http://r4ds.had.co.nz/relational-data.html) framework 
    could be used.
    + The **Serological dataset** and **Covariate/Epidemiological** dataset could 
    keep to be separated but **related**, instead.
    + More details in [APPENDIX](#appendix-tidy-vs-untidy)
- re-evaluate 01 patient 02 id after `unap6`
    + evaluation in [APPENDIX](#appendix-id-errors)



## Dependencies

The required R packages for this analysis are:

+ `tidyverse` [@tidyverse]
+ `haven` [@haven]
+ `readxl` [@readxl]
+ `forcats` [@forcats]
+ `stringr` [@stringr]

```{r, results='hide', message=FALSE}
##essential
library(XLConnect)    # load EXCEL workbooks
library(drc)          # Dose-Response modeling
##accesory
library("DiagrammeR") # method Flowchart
library("knitr")      # To display nice tables
library("tidyverse")
#ggplot2, tibble, tidyr, readr, purr, dplyr
##extra
library("Rmisc")      # multiplot ggplots
library("haven")      # import STATA df
library("readxl")     # import EXCEL df
library("forcats")    # factor vectors
library("stringr")    # for strings
library("viridis")    # color visualization
```

## Input
```{r}
std.nty <- readRDS("data/02-dtfilter-nty.rds")
std.uty <- readRDS("data/02-dtfilter-uty.rds")
```

- total amount of observations **`r length(std.nty$ID)`** in `std.nty`.
- total amount of evaluated patients **`r length(levels(std.nty$ID))`** in `std.nty`.
- giving **`r length(std.uty$ID)`** available samples in `std.uty`.

## Covariates {#bashcov}

The original non-standardized names of the **06** datasets with covariates are:

```{bash}
ls data-raw/ | grep -v "Template" | grep -v "zip" | grep -v "do"
```

In order to follow standardize names, the files we copied and renamed:

```{bash}
cp data-raw/UNAPdatabase20Sept16-pre.dta \
data/UNAP-1-DB-20sep2016.dta

cp data-raw/Zungarococha_database1_3Oct2016.csv \
data/UNAP-2-DB-03oct2016.csv

cp data-raw/UNAPdatabase20Sept16.dta \
data/UNAP-3-DB-02nov2016.dta

cp data-raw/RESULTADOS_LABORATORIO_PRIMER\ BARRIDO\ \(JUL-\ AGO\ 2015\).xlsx \
data/UNAP-4-UP-16ene2017.xlsx

cp data-raw/QC\ results\ Zungarococha.xlsx \
data/UNAP-5-QC-08feb2017.xlsx

cp data-raw/Base\ datos\ epidemiologicos\ 2015.xlsx \
data/UNAP-6-EPI-08feb2017.xlsx
```

New names include:

- order of appearance throughout the project
- relevand info: DB, UP, QC or EPI
- date of registered last edition, or
- receipt date

```{bash}
ls -sh data/UNAP-*
```

## Priors {#priorupgraph}

Prior information before the former analysis:

- `unap1` should be equal to `unap2`
- `unap3` should be equal to `unap2` plus seropositivity agregate columns
- `unap3` was used in the online meeting with Nuno
- `unap4` have updated covariates and differs from `unap3`
- `unap5` have QC evaluations
- `unap6` was the original epidemilogical dataset.

## Updates

Updated information after the analysis was done:

- `unap6` is the **source** for `unap1` age and sex covariates only.
- `unap1` is the **source** of all other `unap#` DB's
- `unap2` is a **subset** of `unap1` shared with Nuno
- `unap3` is `unap1` plus **05** more seropositivity covariates
- `unap4` is `unap1` but with **updated** age and sex covariates
- `unap5` is `unap1` plus **QC** evaluations of micro and pcr covariates
- `unap5` is **different** from `unap1` only in micro UNAP covariate
- `unap1` UNAP covariates were arbitrarily **selected** to compare against `unap5` NAMR ones.

A diagram of DB history using `DiagrammeR` [@DiagrammeR]

```{r, fig.align='center', fig.width=9, fig.height=3}
#install.packages("DiagrammeR")
#library(DiagrammeR)
# node definitions with substituted label text
# edge definitions with the node IDs
DiagrammeR::grViz("
digraph unap_summary {

graph [layout= dot,
        rankdir= LR]


node [fontname = Helvetica]
a [label = '@@1']
b [label = '@@2']
c [label = '@@3']
d [label = '@@4', color= red, fontcolor= red]
e [label = '@@5', color= red, fontcolor= red]
f [label = '@@6', color= green, fontcolor= green]
g [label = '@@7', color= blue, fontcolor= blue]


a -> {b c}
a -> {d e}  [color= red]
f -> a [color= green]
{a d e} -> g [color= blue]
}

[1]: 'unap1 DB (27)'
[2]: 'unap2 DB (8)'
[3]: 'unap3 DB (32)'
[4]: 'unap4 UP (8)'
[5]: 'unap5 QC (4)'
[6]: 'unap6 EPI (109)'
[7]: 'unap0 DB (14)'
")
```


## Procedure

### UNAP-1-DB-20sep2016.dta

discoveries:

- `unap1` already have part of the seropositivity output (also present in `unap3`)
- variables absent in `unap2` are listed
- removed columns:
    + micro2 and dens2
    + OD reads
    + sero-positive results (cutoff + categories)
    + age categories
    + pcrsnou 1/2 and pcrrubio 1/2

```{r}
unap1 <- haven::read_dta("data/UNAP-1-DB-20sep2016.dta") %>%
  haven::as_factor() #%>% dplyr::select(id,contains("pcr")) %>% mutate(pcrsnou1=as.factor(pcrsnou1),pcrsnou2=as.factor(pcrsnou2)) %>% 
  #dplyr::count(pcrrubio1,pcrrubio2,pcrsnou1,pcrsnou2,pcrrubio)#summary()
# previous removing
unap1.out <- unap1 %>% select(micro2:age3, -pcrrubio)
unap1 <- unap1 %>% select(-(micro2:age3), pcrrubio)
#unap1$id <- forcats::as_factor(unap1$id) # not required
str(unap1)
summary(unap1)
```

```{r}
colnames(unap1.out)
```


### UNAP-2-DB-03oct2016.csv

discoveries:

- `unap2` is a **subset** of `unap1`
- `unap2` do not used **pvmsp1n** variable
- removed columns:
    + OD reads


```{r}
unap2 <- readr::read_csv("data/UNAP-2-DB-03oct2016.csv") %>%
  dplyr::mutate_each(funs(factor), 2, 4, 5, 9) #%>% glimpse()
#str(unap2)
unap2.out <- unap2 %>% select(matches(".msp.")) # previous removing
unap2 <- unap2 %>% select(-matches(".msp.")) # OD reads removing
str(unap2)
summary(unap2)
```
```{r}
colnames(unap2.out)
```



### UNAP-3-DB-02nov2016.dta

discoveries:

- `unap3` is identical to `unap1` plus **05 more columns** about seropositivity
- removed columns:
    + micro2 and dens2
    + OD reads
    + sero-positive results (cutoff + categories + 5 columns)
    + age categories
    + pcrsnou 1/2 and pcrrubio 1/2

```{r}
unap3 <- haven::read_dta("data/UNAP-3-DB-02nov2016.dta") %>% 
  haven::as_factor() #%>% glimpse()
#str(unap3)
# previous removing
unap3.out <- unap3 %>% select(micro2:ll_pvmsp1pos, -pcrrubio)
unap3 <- unap3 %>% select(-(micro2:ll_pvmsp1pos), pcrrubio)
#unap3$id <- forcats::as_factor(unap3$id) # not required
str(unap3)
summary(unap3)
```

```{r}
colnames(unap3.out)
#summary(select(unap3.out, (seroposall:ll_pvmsp1pos)))
```


### UNAP-4-UP-16ene2017.xlsx

discoveries:

- `unap4` is identical to `unap1` with **updated** `age` and `sex` covariates
- `unap1` is the original source of all `unap#` covariates
    + then, all covariates should be retrieved from `unap1`

```{r}
#read xl
unap4 <- readxl::read_excel("data/UNAP-4-UP-16ene2017.xlsx", 
                            sheet = 1, na = ".") %>%
  dplyr::select(id=1, sex=3, age=2, community=4, micro1=5, dens1=6, 
                pcrrubio_Pv=12, pcrrubio_Pf=13, pcrsnounou_Pv=14, pcrsnounou_Pf=15) %>%
  dplyr::mutate(pcrrubio= pcrrubio_Pf + pcrrubio_Pv,
                pcrsnounou= pcrsnounou_Pf + pcrsnounou_Pv) %>%
  dplyr::select(id:dens1,pcrrubio, pcrsnounou) %>%
  dplyr::mutate_each(funs(factor), 2,4,5,7,8)
str(unap4)
summary(unap4)
```

#### unap4 covariate labels

Covariate labeling is based on labels available in covariate DB sheet 2.

```{r}
unap4.xl <- XLConnect::loadWorkbook("data/UNAP-4-UP-16ene2017.xlsx")
#
unap4.lb <- XLConnect::readWorksheet(unap4.xl, header = F,
                      sheet = 2, 
                      startRow = 8, endRow = 16,
                      startCol = 2, endCol = 6)
unap4.lb
```


### UNAP-5-QC-08feb2017.xlsx

discoveries:

- `unap1` **micro1** is different from `unap5` **micro.unap**
- **why?**

```{r}
#read xl
unap5 <- readxl::read_excel("data/UNAP-5-QC-08feb2017.xlsx", na = "-") %>% 
  select(1:5) %>% 
#colnames(unap5)
#unap5 <- unap5[,1:5]
#
#unap5 <- unap5 %>% 
  dplyr::select(id=1, pcr.unap=4, pcr.namr=5,  
                micro.unap=2, micro.namr=3) %>%
  dplyr::mutate(micro.comp= micro.unap==micro.namr,
                pcr.comp= pcr.unap==pcr.namr) %>%
  dplyr::select(1:3,7,4:5,6) #%>%
  #dplyr::mutate_each(funs(factor), 2,3,5,6)
str(unap5)
summary(unap5 %>%
          select(1,4,7))
```

#### test equality

- `unap1` and `unap5` should have equal **micro** and **pcr** ***UNAP*** covariates
- suffix:
    + `_1`: `unap1`
    + `_5`: `unap5` **UNAP** covariates
    + `_0`: comparison between `_1` and `_5`
    + `_n`: `unap5` **NAMR** covariates
    + `_n1`: `_n` vs `_1`
    + `_n5`: `_n` vs `_5`

```{r}
#unap5 <- readxl::read_excel("data/UNAP-5-QC-08feb2017.xlsx", na = "-")
unap5c <- unap5 %>%
  select(id, micro_5= micro.unap, pcr_5=pcr.unap) %>%
  join(unap1 %>% #full_join(unap1 %>%
              select(id, micro_1= micro1, pcr_1=pcrrubio), 
            by= "id") %>%
  mutate(micro_1= fct_recode(micro_1,
                             "0"= "Negative", 
                             "1"= "P vivax", 
                             "2"= "P falciparum", 
                             "3"= "Pv/Pf mixed infection"),
         pcr_1= fct_recode(pcr_1, 
                           "0"= "Negative", 
                           "1"= "P vivax", 
                           "2"= "P falciparum",
                           "3"= "Pv/Pf mixed infection")) %>%
  mutate(micro_0= micro_1==micro_5, 
         pcr_0= pcr_1==pcr_5) %>%
  full_join(unap5 %>%
              select(id, micro_n= micro.namr, pcr_n= pcr.namr), 
            by="id") %>%
  mutate(micro_n1= micro_n==micro_1,
         micro_n5= micro_n==micro_5,
         pcr_n1= pcr_n==pcr_1,
         pcr_n5= pcr_n==pcr_5) %>%
  select(1,4,2,6,8,10,11,5,3,7,9,12,13) %>%
  dplyr::mutate_each(funs(factor), 3,5,9,11)
  
str(unap5c)
```

**pcr_1** and **pcr_5** are equivalent:

```{r}
summary(unap5c %>% 
          select(8:13))
```

However, **micro_1** and **micro_5** are different in **43** cells

```{r}
summary(unap5c %>% 
          select(2:7))
```

#### selection

Arbitrary selection of `unap1` as **UNAP** covariates to compare against `unap5` **NAMRU** ones.

```{r}
unap5d <- unap5c %>%
  select(1, 
         2,5,6, 
         8,11,12)
summary(unap5d)
```


### UNAP-6-EPI-08feb2017.xlsx

- New question, New .Rmd

```{r}
#read xl
unap6 <- readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 1) #%>% glimpse()
#filter NA
unap6 <- unap6[1:sum(unap6$id_integrante != "", na.rm = T), 
           1:sum(colnames(unap6) != "", na.rm= T)]
#963x109
#colnames(unap6)
#str(unap6)
```

## Merge covariates

AIM:

- create `unap0`
- merge 
    + `unap1` age, sex
    + `unap4` age, sex
    + change factor label in sex
    + create 2 logic variables
    + `unap5` all (micro, pcr)
    + `unap1` community
    + avoided: dens1, micro2, dens2, pcr.snounou
- lastly, 
    + create a concensus for `micro_` and `pcr_`
- suffix:
    + `_1`: `unap1`
    + `_4`: `unap4`
    + `_0`: comparison between `_1` and `_4`
    + `_n`: `unap5` **NAMR** covariates
    + `_n1`: `_n` vs `_1`
    + `_c`: consensus variable between `_n` vs `_1`

A plot showing How `age` updating changed with respect to the patient `Ab.units` is in the 
[APPENDIX](#appendix-age-vs-ab-unit)

```{r}
#str(unap1)
#str(unap4)
unap4$sex <- fct_recode(unap4$sex, Female="1", Male="2")
#str(unap4)

unap0 <- 
  full_join(
    unap1 %>%
      select(id, age, sex), 
    unap4 %>%
      select(id, age, sex), 
    by= "id", suffix= c("_1", "_4")) %>%
  mutate(sex_0= sex_1==sex_4,
         age_0= age_1==age_4)%>%
  select(1,3,5,6,2,4,7)%>%
  full_join(unap5d, by= "id") %>%
  full_join(unap1 %>%
              select(id,community_1= community#, 
                     #micro_1= micro1, dens_1= dens1, pcr_1=pcrrubio
                     ),
            by="id")
```

### Generate consensus

- `micro` and `pcr` consensus variables
- criteria:
    + NAMR measurement had preference

```{r}
unap0 <- unap0 %>% 
  #select(contains("micro")) %>% #, -micro_n1
  tidyr::unite(micro_c, micro_1, micro_n,remove = F) %>% #count()
  mutate(micro_c=str_replace(micro_c,"(\\d)_\\1", "\\1")) %>%
  mutate(micro_c=str_replace(micro_c,"(.+)_NA", "\\1")) %>% 
  mutate(micro_c=str_replace(micro_c,"(\\d)_(\\d)", "\\2")) %>%  #count()
  mutate(micro_c=str_replace(micro_c,"NA", replacement = NA_character_)) %>%  #%>% count()
  #select(contains("pcr")) %>% #, -micro_n1
  tidyr::unite(pcr_c, pcr_1, pcr_n,remove = F) %>% #count()
  mutate(pcr_c=str_replace(pcr_c,"(\\d)_\\1", "\\1")) %>%
  mutate(pcr_c=str_replace(pcr_c,"(.+)_NA", "\\1")) %>%   #count()
  mutate(pcr_c=str_replace(pcr_c,"(\\d)_(\\d)", "\\2")) %>% #count() #summary()
  mutate(pcr_c=str_replace(pcr_c,"NA", replacement = NA_character_)) %>% dplyr::mutate_each(funs(factor),8,12)

#
#unap0
str(unap0)
#summary(unap0)
```



## Merge Ab units

- `std.uty` can be directly merged with `unap0`
- `std.nty` + `unap0` assumes that:
    + `micro_` and `pcr_` are independent observations with regard to the `specie`.

### untidy format

- At the moment, the most convenient dataset.
- Check its tidy version summary [here](#validate-tidy)

```{r}
std.uty.cov <- std.uty %>%
  select(id=ID, 2:5) %>%
  full_join(unap0, by="id") #%>%
  #select(1:14,21,15:17,19,20,18)


str(std.uty.cov)
summary(std.uty.cov)
#View(std.uty.cov)
```

### tidy format

- `std.nty` + `unap0` assumes that:
    + `micro_` and `pcr_` are independent observations with regard to the `specie`.

**Would require:**

- merge `std.nty` and `unap0`
- filter by `Specie`
- recode `micro_` and `pcr_` levels per `Specie`
- generate `std.nty.cov` in a tidy format by combining all cases per Specie
- correct `micro_n1` and `pcr_n1` logic vectors

**Instead:**

- A [Relational data](http://r4ds.had.co.nz/relational-data.html) framework could be used.
    + The **Serological dataset** and **Covariate/Epidemiological** dataset would be 
    separated but **related**.
- This would imply:
    + keep `std.nty` as it is: tidy, and avoid `std.uty`.
    + keep `unap0` as it is: with no issues that came from the `Specie` covariate.
    + avoid untiding `std.nty` + `unap0` for `micro_` and `pcr_` coherence.
- At the moment, the most convenient dataset is `std.uty.cov`.

**By now:**

- `micro_` and `pcr_` are `Specie` related covariates
- `micro_` and `pcr_` were merged with serological data.

**TO DO:**

- A concensus, as in untidy, is missing.

```{r}
std.nty.pre <- std.nty %>%
  select(id=ID, 4:6) %>%
  full_join(unap0, by="id") #%>% count("micro_1")

std.nty.pre_Pv <- std.nty.pre %>% 
  select(1:4, micro_1:pcr_n1, -pcr_c) %>% 
  filter(Specie=="Pviv") %>% #count("pcr_1")
  mutate(micro_1=fct_recode(micro_1,
                            "0"="2",
                            "1"="3"),
         micro_n=fct_recode(micro_n,
                            "0"="2",
                            "1"="3"),
         pcr_1=fct_recode(pcr_1,
                          "0"="2"),
         pcr_n=fct_recode(pcr_n,
                          "0"="2",
                          "1"="3")) #%>% count("pcr_n")

std.nty.pre_Pf <- std.nty.pre %>% 
  select(1:4, micro_1:pcr_n1, -pcr_c) %>% 
  filter(Specie=="Pfal") %>% #count("pcr_n")
  mutate(micro_1=fct_recode(micro_1,
                            "1"="2",
                            "0"="1",
                            "1"="3"),
         micro_n=fct_recode(micro_n,
                            "1"="2",
                            "0"="1",
                            "1"="3"),
         pcr_1=fct_recode(pcr_1,
                          "1"="2",
                          "0"="1"),
         pcr_n=fct_recode(pcr_n,
                          "1"="2",
                          "0"="1",
                          "1"="3")) #%>% count("micro_1")

std.nty.cov <- dplyr::union(std.nty.pre_Pf, std.nty.pre_Pv)

## correct logic vectors of micro and pcr
std.nty.cov <- std.nty.cov %>% 
  mutate(micro_n1= micro_1==micro_n,
         pcr_n1= pcr_1==pcr_n)


str(std.nty.cov)
#summary(std.nty.cov)
```

**CHECK:**

- Note that `micro_` and `pcr_` covariates are now standardized as factors with only **2** labels.
- Validation in [APPENDIX](#appendix-tidy-vs-untidy)



## QC report

### using untidy dataset
```{r}
std.uty.cov %>% 
  select(micro_n1) %>% 
  filter(micro_n1!="NA") %>% 
  count() %>% 
  mutate(percent_similarity= (freq/sum(freq))*100)
  
std.uty.cov %>% 
  select(pcr_n1) %>% 
  filter(pcr_n1!="NA") %>% 
  count() %>% 
  mutate(percent_similarity= (freq/sum(freq))*100)
```

### using tidy dataset

- assumes that `micro_` and `pcr_` are independent observations with regard to the `specie`.
- given the mentioned assumption, percentage increases due to an updated logic variables.
    + Check details in [APPENDIX](#appendix-tidy-vs-untidy)

```{r}
std.nty.cov %>% 
  select(micro_n1) %>% 
  filter(micro_n1!="NA") %>% 
  count() %>% 
  mutate(percent_similarity= (freq/sum(freq))*100)

std.nty.cov %>% 
  select(pcr_n1) %>% 
  filter(pcr_n1!="NA") %>% 
  count() %>% 
  mutate(percent_similarity= (freq/sum(freq))*100)
```


## Output

- `std.uty.cov` output was written as `data/03-covariat-uty.csv`.
- `std.nty.cov` output was written as `data/03-covariat-nty.csv`.

```{r}
#
saveRDS(unap0, "data/03-covariat-uap.rds")
#
#readr::write_csv(std.nty.cov %>% format(scientific=FALSE, digits=2), "data/03-covariat-nty.csv")
saveRDS(std.nty.cov, "data/03-covariat-nty.rds")
#
#readr::write_csv(std.uty.cov %>% format(scientific=FALSE, digits=2), "data/03-covariat-uty.csv")
saveRDS(std.uty.cov, "data/03-covariat-uty.rds")
```


***

## APPENDIX: age vs Ab unit

- the difference between `age_1` and `age_4` was stored in `age_d` and plotted against `Ab.units`.

```{r, fig.align='center', fig.width=8, fig.height=5}
db_ageAb <- std.uty.cov %>% 
  select(1:5,contains("age")) %>% 
  filter(age_0==FALSE) %>% 
  mutate(age_d= age_4-age_1) 

a <- db_ageAb %>% 
  ggplot(aes(Ab.unit_Pviv,age_d)) +
  geom_point()+
  geom_hline(aes(yintercept=0), colour="red")

b <- db_ageAb %>% 
  ggplot(aes(Ab.unit_Pfal,age_d)) +
  geom_point()+
  geom_hline(aes(yintercept=0), colour="red")

c <- db_ageAb %>% 
  ggplot(aes(mean.OD_Pviv,age_d)) +
  geom_point()+
  geom_hline(aes(yintercept=0), colour="red")

d <- db_ageAb %>% 
  ggplot(aes(mean.OD_Pfal,age_d)) +
  geom_point()+
  geom_hline(aes(yintercept=0), colour="red")

Rmisc::multiplot(c,d,a,b,cols = 2)
```


## APPENDIX: tidy vs untidy

### How to subset mixed infections with tidy?

- with regard to `pcr_n`

```{r}
std.nty.cov %>% 
  select(1,4,9) %>% #count("pcr_n")
  mutate(pcr_n= as.numeric(pcr_n)) %>% #count("pcr_n") # WARNING: 0 and 1 -> 1 and 2
  group_by(id) %>%
  summarise_if(is.numeric,sum) %>% # count("pcr_n")
  filter(pcr_n=="4") %>% 
  select(id)
  #filter(id=="LL013-1")#4 = 2+2 = mixed infection
  #filter(id=="LL011-3")#3 = 2+1 | 1+2 = mono infections
  #filter(id=="LL022-3")#2 = 1+1 = no infection
  
```


### Validate tidy

- Contrast this subset against `summary(std.uty.cov)` [here](#untidy-format):

```{r}
# ONLY FOR SEROLOGICAL DATA and -if required- for MICRO AND PCR data
summary(std.nty.cov %>% filter(Specie=="Pfal"))
summary(std.nty.cov %>% filter(Specie=="Pviv"))
```

- updated logic variables in `std.nty.cov`

```{r}
## TIDY DF needs NEW LOGIC vectors
std.nty.cov %>% filter(id=="ZG178-5") %>% select(-3)
std.uty.cov %>% filter(id=="ZG178-5") %>% select(1:3,12:17)
```

### tidy is relevant for the QC evaluation {#tidyqcrev}
- In case **NAMR** differs in the identification of mixed-infection with respect to **UNAP**, 
- tidy dataset **QC report** takes **UNAP** mono-infection detections into account.
```{r}
# Pvivax + Pfalcip
std.nty.cov %>% 
  #select(1,2,4,11:16) %>% 
  filter(id=="LL013-1") # problem solved!!
std.uty.cov %>% 
  select(1:3,12:17) %>% 
  filter(id=="LL013-1")
#unap5 %>% 
#  filter(id=="LL013-1")
```

```{r, eval=FALSE}
## TEST all cases

# Pvivax
std.nty.cov %>% 
  select(1:4,11:16) %>% 
  filter(id=="LL003-4") # ok
# Pfalcip
std.nty.cov %>% 
  select(1:4,11:16) %>% 
  filter(id=="LL038-3") # ok
# non
std.nty.cov %>% 
  select(1:4,11:16) %>% 
  filter(id=="LL008-8") # ok
```




## APPENDIX: ID errors

### EVALUATE 01 patient 02 IDs

- TO DO:
    + In stand-by until finished with the epidemiological database `unap6`.

- AIM:
    + Contrast the data of the samples mentioned in the 2nd sheet of `UNAP-4-UP-16ene2017.xlsx`

```{r}
#unap4.xl <- XLConnect::loadWorkbook("data/UNAP-4-UP-16ene2017.xlsx")
unap4.dl <- readWorksheet(unap4.xl, 
                      sheet = 2, 
                      startRow = 20, endRow = 27,
                      startCol = 2, endCol = 3)
unap4.dl

#str(unap4.dl)
```

Strings were split into individual characters in the following way:
```{r}
strsplit(unap4.dl[5,], split = " ")[[1]]
```

Check equality between pair of samples:
```{r, eval=FALSE}
unap4.dlMIX <- data.frame()
for(i in 1:dim(unap4.dl)[1]) {
  unap4.dlMIX <- rbind(unap4.dlMIX,
                      std.uty.cov[std.uty.cov$id==
                                   strsplit(unap4.dl[i,], 
                                            split = " ")[[1]][1],],
                      std.uty.cov[std.uty.cov$id==
                                   strsplit(unap4.dl[i,], 
                                            split = " ")[[1]][3],])
}
unap4.dlMIX
#format(unap4.dlMIX[,-c(7,8,10)], scientific=FALSE, digits=2)#
```

Some equalities do not seem to be from the same patient. This step needs more support prior to any sample deletion.


***

<!--## References-->
