==> 04-epidemio.Rmd <==
# EPIDEMIOLOGICAL COVARIATES {#epide}

## Aim

* Principal: 
    - Clean and standardize the UNAP-6 Epidemiological and Census database.

* Secondary: 
    - Check variable consistency across the database.

## Results

- The **109 variables** of the Epidemiological database were reduced to **88** 
after cleaning.
- typos were corrected:
    + some variables were renamed, and some factors recoded. [here](#typo-correction)
- previous datasets share **959 ID's** with the current Epidemiological dataset. [here](#merge-datasets)
- Two unified dataset were generated:
    + One with the **full 973** ID's
    + One with only the **core 959** ID's
- a plot `age` vs `residence` show some inconsistencies corrected throughout updates. [here](#age-vs-residence)
- **inconsistencies** between other variables are shown as table summaries. [here](#problemasx)
- `sheet2` included as **[Data Dictionary](#data-dictionary)**. [here](#data-dictionary)
- file output as a `.dta`. [here](#write-dta)


## To Do

- Solve **inconsistencies**
- CHECK IF **intersections** outside the core **959 samples** have ID typos.
- introduce a `datetime` class variable


## Dependencies

The required R packages for this analysis are:

+ `tidyverse` [@tidyverse]
+ `readxl` [@readxl]
+ `forcats` [@forcats]
+ `stringr` [@stringr]

```{r, results='hide', message=FALSE}
##essential
library(XLConnect)    # load EXCEL workbooks
library(drc)          # Dose-Response modeling
##accesory
library("DiagrammeR") # method Flowchart
library("knitr")      # To display nice tables
library("tidyverse")
#ggplot2, tibble, tidyr, readr, purr, dplyyr
##extra
library("Rmisc")      # multiplot ggplots
library("haven")      # import STATA df
library("readxl")     # import EXCEL df
library("forcats")    # factor vectors
library("lubridate")  # for dates
library("stringr")    # for strings
library("viridis")    # color visualization
```

## Input
```{r}
std.nty <- readRDS("data/02-dtfilter-nty.rds")
std.uty <- readRDS("data/02-dtfilter-uty.rds")

unap0 <- readRDS("data/03-covariat-uap.rds")
std.nty.cov <- readRDS("data/03-covariat-nty.rds")
std.uty.cov <- readRDS("data/03-covariat-uty.rds")
```


## Covariate datasets

```{bash}
ls -sh data/UNAP-*
```


## UNAP-6-EPI-08feb2017.xlsx
```{r}
#check sheets
readxl::excel_sheets("data/UNAP-6-EPI-08feb2017.xlsx")
```

### sheet 1
```{r}
#read xl
#read xl
unap6 <- readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 1) %>% 
  mutate(anios_residencia= as.numeric(anios_residencia))
#filter NA
unap6 <- unap6[1:sum(unap6$id_integrante != "", na.rm = T), 
           1:sum(colnames(unap6) != "", na.rm= T)]
#963x109
unap6 %>% glimpse()
#colnames(unap6)
#str(unap6)
#summary(unap6)
```

### typo correction

- rename variables
- recode factors
- `unap6a` created

Bad idea:

- remove varaibles with full **NA** values (109 -> 95)
- affects class of numeric variables

TO DO:

- who to fix it?
- function to identify numeric values inside characters: **`chr` -> `dbl`**

```{r}
unap6a <- unap6 %>%
  dplyr::rename(id=id_integrante, 
         community=barrio, 
         age= Edad,
         residence= anios_residencia,
         sex= sexo,
         trabajo_paga_ultima= trabajo_paga_utlima,
         banio_conectado= banio_contectado,
         drogas_tratamiento_01= drogras_tratamiento_01,
         ndias_drogas_tratamiento_01= ndias_drogras_tratamiento_01,
         drogas_tratamiento_02= drogras_tratamiento_02,
         ndias_drogas_tratamiento_02= ndias_drogras_tratamiento_02,
         drogas_tratamiento_03= drogras_tratamiento_03,
         ndias_drogas_tratamiento_03= ndias_drogras_tratamiento_03) %>%
  mutate(id= str_replace_all(id, "_","-"),
         sex= fct_recode(sex, M="m"), 
         nivel_educacion= fct_recode(nivel_educacion, 
                                    "Tecnica Completa"="TÃƒÂ©cnica Completa",
                                    "Tecnica Incompleta"="TÃƒÂ©cnica Incompleta"),
         otro_tipo_trabajo= fct_recode(otro_tipo_trabajo, 
                                    "albanil"="albaÃƒÂ±il",
                                    "albanil"="albaÃ±il",
                                    "contruccion"="construcciÃƒÂ³n"),
         otro_material_paredes= fct_recode(otro_material_paredes, 
                                    "Tripley"="tripley",
                                    "Tripley"="tripleyy"),
         fuente_agua= fct_recode(fuente_agua,
                                "3. Tuberia publica, fuente fuera de la casa, 
                                compartido con la comunidad"=
                                  "3. TuberiƒÂa publica, fuente fuera de la casa, 
                                compartido con la comunidad"),
         banio_conectado= fct_recode(banio_conectado,
                                     "0. No tiene banio"="0. No tiene baÃ±o"),
         recibio_tratamiento_malaria= fct_recode(recibio_tratamiento_malaria,
                                                "1. Mas de una vez (especifique)"=
                                                  "1.Mas de una vez (especifique)",
                                                "1. Mas de una vez (especifique)"=
                                                  "1. MAs de una vez (especifique)",
                                                "1. Mas de una vez (especifique)"=
                                                  "1. MÃ¡s de una vez (especifique)"),
         epi_alguien_tuvo_malaria= fct_recode(epi_alguien_tuvo_malaria,
                                             "2. Si, hace algunos anos"=
                                               "2. Si, hace algunos aÃ±os",
                                             "3. Si, el ultimo ano"=
                                               "3. Si, el ultimo aÃ±o"),
         no_busco_ayuda_tratamiento= fct_recode(no_busco_ayuda_tratamiento,
                                               "0. No pensaba que era un problema serio"=
                                                 "0. No pensa que era un problema serio",
                                               "0. No pensaba que era un problema serio"=
                                                 "0. No pensÃƒÆ’Ã‚Â³ que era un problema serio"),
         drogas_tratamiento_01= fct_recode(drogas_tratamiento_01,
                                           "artesunato"="artesunatu",
                                           "artesunato"="Artesunatu",
                                           "cloroquina"="Cloroquina ",
                                           "cloroquina"="Cloroquina",
                                           "FANSIDAR"="fansedal",
                                           "primaquina"="trimaquina"),
         drogas_tratamiento_02= fct_recode(drogas_tratamiento_02,
                                           "clindamicina"="Clinamicina",
                                           "mefloquina"="nefloquina",
                                           "mefloquina"="Nefloquina",
                                           "primaquina"="Pramaquina",
                                           "primaquina"="Primaquima",
                                           "primaquina"="Primaquina",
                                           "primaquina"="trimaquina"),
         enf_especifique_renal=fct_recode(enf_especifique_renal,
                                          "dolor de rinones"="dolor de riÃ±oned",
                                          "dolor de rinones"="dolor de RiÃ±ones",
                                          "dolor de rinones"="dolores",
                                          "piedras en rinones"="piedras en los riÃ±ones",
                                          "piedras en rinones"="piedrillas"),
         enf_especifique_otro=fct_recode(enf_especifique_otro,
                                         "gastritis"="Gastritis",
                                         "migrana"="Migrana"),
         enf_medicinas_parasitos=fct_recode(enf_medicinas_parasitos,
                                         "albendazol"="abbendazol",
                                         "albendazol"="abendabzol",
                                         "albendazol"="abendazol",
                                         "albendazol"="Abendazol",
                                         "albendazol"="abendazol3",
                                         "albendazol"="abendazolp",
                                         "albendazol"="abenzadol",
                                         "albendazol"="albendazol",
                                         "antiparasitario"="antiparasitarias",
                                         "antiparasitario"="antiparasitarios"),
         combustible_cocinar=fct_recode(combustible_cocinar,
                                        "5. Lenha"="5. Leña")
         ) #%>% mutate_each(funs(factor), 3) 
  #gather(key= variable, value = observation, -id, na.rm = T) %>% 
  #spread(variable, observation) 

#unap6a %>% glimpse()

#unap6a %>% #dplyr::count(epi_alguien_tuvo_malaria, sort = T)
#  str()
  #count(tipo_trabajo)
  #count(combustible_cocinar)
  #count(material_paredes)
#count(material_piso)
  #count(fuente_agua)
  #count(recibio_tratamiento_malaria)
  #count(drogras_tratamiento_01) #--------------------dirty
  #count(drogras_tratamiento_02) #--------------------dirty
  #count(fecha_registro) #--------------------------------------lubridate
  #count(epi_especie_causo_malaria) %>%
  #mutate(prop=n/sum(n))
```

#### all NA variables

- **14** variables are full of NA values
- remove throughout the evaluations

```{r}
unap6x <- unap6a %>%   
  gather(key= variable, value = observation, -id, na.rm = T) %>% 
  spread(variable, observation)

xd <- data_frame(v= colnames(unap6x))
ad <- data_frame(v= colnames(unap6a))

dplyr::setdiff(ad,xd)
```


### problems 1

#### age vs residence

discovery:

- `unap6` differs from `unap1` in **age** variable

TO DO:

- CHECK WHY some `age_4` are still lower than patient `residence` time.

```{r ageres, fig.width=12, fig.height=3, fig.align="center"}
unap60ar <- std.uty.cov %>% 
  dplyr::select(id, age_1, age_4) %>% 
  left_join(unap6a %>% 
              dplyr::select(id, age_e=age, residence), by="id") %>% 
  mutate(age_e1= age_1==age_e,
         age_e4= age_4==age_e,
         age_0= age_1==age_4) #%>% #filter(age_e1==F) #count("age_e1")

a <- unap60ar %>% 
  group_by(age_e, residence) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(age_e, residence, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="unap6-EPI")
  
b <- unap60ar %>% 
  group_by(age_1, residence) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(age_1, residence, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="unap1-DB")

c <- unap60ar %>% 
  group_by(age_4, residence) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(age_4, residence, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="unap4-UP")

Rmisc::multiplot(a,b,c,cols = 3)
```

### problems 2 {#problemasx}

- `unap6a` was updated on each step

#### ah_

- change `ah_` factor labels

```{r}
unap6a <- unap6a %>% 
  mutate_at(vars(ah_radio:ah_auto), factor) %>% #dplyr::select(1, dplyr::contains("enf")) %>% glimpse()
  mutate(ah_radio=fct_recode(ah_radio,"si"="1","no"="0"), 
         ah_television=fct_recode(ah_television,"si"="1","no"="0"), 
         ah_estereo=fct_recode(ah_estereo,"si"="1","no"="0"), 
         ah_refrigerador=fct_recode(ah_refrigerador,"si"="1","no"="0"), 
         ah_motocicleta=fct_recode(ah_motocicleta,"si"="1","no"="0"), 
         ah_mototaxi=fct_recode(ah_mototaxi,"si"="1","no"="0"), 
         ah_auto=fct_recode(ah_auto,"si"="1","no"="0")) #%>% dplyr::select(enf_diabetes_si,enf_diabetes_no) %>% count()

unap6a %>% 
  dplyr::select(1, starts_with("ah_")) %>%
  gather(key= ah, value = ah_m, -id, na.rm = T) %>% 
  filter(ah_m=="si") %>% dplyr::count(ah, sort = T)
  #group_by(ah, ah_m) %>% 
  #dplyr::summarise(n=n()) %>%
  #mutate(prop= (n/sum(n)*100))
```

#### otro_

discovery:

- this step informs that only **3 out of 11** variables that start with `otro_` have significant variables.
- remove **8** `otro_` variables

```{r}
unap6a %>% 
  dplyr::select(1, starts_with("otro_")) %>% #str() # 3/11 son significativas
  gather(key= otro, value = otro_m, -id, na.rm = T) %>% dplyr::count(otro, sort = T)

unap6_otro <- unap6a %>% 
  dplyr::select(1, starts_with("otro_")) %>% #str() # 3/11 son significativas
  dplyr::select(1, 2,6,9) #%>% #str() # 3/11 son significativas
  
unap6a <- unap6a %>% 
  dplyr::select(-starts_with("otro_")) %>% #str() # 3/11 son significativas
  left_join(unap6_otro, by="id") #%>% dplyr::select(1, starts_with("otro_")) %>% str() # 3/11 son significativas
  #spread(key= otro, value = otro_m)
```

#### trabajo

- create `trabajo` in order to merge `tipo_trabajo`and `otro_tipo_trabajo`
- `trabajo` is a **lumped** `fct_lump()` variable

```{r}
unap6a <- unap6a %>% 
  mutate(tipo_trabajo= str_replace(tipo_trabajo,"\\d+\\.\\s", "")) %>% #count("tipo_trabajo")
  mutate(tipo_trabajo= str_replace(tipo_trabajo,"Actualmente ", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(tipo_trabajo= str_replace(tipo_trabajo,str_c("Otro ","\\(","especificar","\\)"), "NA")) %>% #dplyr::count(tipo_trabajo, sort = T)
  unite(trabajo, tipo_trabajo, otro_tipo_trabajo) %>% 
  mutate(trabajo= str_replace(trabajo,"_NA", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(trabajo= str_replace(trabajo,"NA_", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(trabajo= str_replace(trabajo,"_", " ")) %>% 
  mutate(trabajo= str_replace(trabajo,"NA", NA_character_)) %>% 
  mutate(trabajo= str_replace(trabajo,"(.+)(\\s)casa(\\s)(.+)|(.+)(\\s)casa|carpintero|Comerciante|panadero|(.+)(\\s)salud|administrativo(\\s)(.+)|limpieza|mecanico|Mecanico|pensionista|personal(\\s)(.+)|Profesor|tecnico(\\s)(.+)", "indoor")) %>% 
  mutate(trabajo= str_replace(trabajo,"Agricultor|obrero(\\s)(.+)|obrero|pescador|vigilante|Vigilante|albanil|mototaxista|Chofer|construccion|contruccion|Ganadero|Granjero|guia(\\s)(.+)|obreri", " outdoor")) #%>% dplyr::count(trabajo, sort = T)
  #mutate(trabajo= fct_lump(trabajo, 5)) 

#unap6a %>% glimpse()

unap6a %>% 
  dplyr::select(1, dplyr::contains("trabajo")) %>% 
  dplyr::count(trabajo, sort = T)
  #group_by(trabajo, trabajo_paga_ultima) %>% dplyr::count(trabajo_paga_ultima, sort = T)
```

#### material

- create `material_pared` in order to merge `material_paredes`and `otro_material_paredes`

```{r}
unap6a <- unap6a %>% 
  #dplyr::select(1, dplyr::contains("material")) %>% 
  mutate(material_piso= str_replace(material_piso,"\\d+\\.\\s", "")) %>% #dplyr::count(material_piso, sort = T)
  mutate(material_paredes= str_replace(material_paredes,"\\d+\\.\\s", "")) %>% #dplyr::count(material_paredes, sort = T)
  mutate(material_paredes= str_replace(material_paredes,"\\so\\s", "_")) %>% #dplyr::count(material_paredes, sort = T)
  mutate(material_paredes= str_replace(material_paredes,str_c("Otro ","\\(","especifique","\\)"), "NA")) %>%#dplyr::count(material_paredes, sort = T)
  #group_by(material_paredes, otro_material_paredes) %>% dplyr::count()
  unite(material_pared, material_paredes, otro_material_paredes) %>% 
  mutate(material_pared= str_replace(material_pared,"_NA", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(material_pared= str_replace(material_pared,"NA_", "")) %>% #dplyr::count(material_pared, sort = T)
  mutate(material_pared= str_replace(material_pared,"Madera(.+)","Madera")) %>% 
  mutate(material_pared= str_replace(material_pared,"(.+)_cemento_(.+)|(.+)_cemento","cemento")) #%>% dplyr::count(material_pared, sort = T)

#unap6a %>% glimpse()

unap6a %>% 
  dplyr::count(material_pared, sort = T)
  #mutate(material_pared= fct_lump(material_pared, 3)) %>% dplyr::count(material_pared, sort = T)
  #group_by(trabajo, trabajo_paga_ultima) %>% dplyr::count(trabajo_paga_ultima, sort = T)
  
unap6a <- unap6a %>% 
  #dplyr::select(1, dplyr::contains("material")) %>% 
  mutate(material_piso= str_replace(material_piso,"\\d+\\.\\s", "")) %>% 
  mutate(material_piso= str_replace(material_piso,"(.+)madera(.+)|(.+)(\\s)madera","madera")) #%>% 
  #mutate(material_piso= str_replace(material_piso,"(.+)_cemento_(.+)|(.+)_cemento","cemento")) #%>% 
  #dplyr::count(material_piso, sort = T)

unap6a %>% dplyr::count(material_piso, sort = T)
```

#### banio

- create `banio_conexion` in order to merge `banio_conectado`and `otro_banio_conectado`

```{r}
unap6a <- unap6a %>% 
  #dplyr::select(1, dplyr::contains("banio")) %>% #dplyr::count(banio_conectado, sort = T)
  mutate(banio_conectado= str_replace(banio_conectado,"\\d+\\.\\s", "")) %>% #dplyr::count(material_paredes, sort = T)
  mutate(banio_conectado= str_replace(banio_conectado,str_c("Otro","\\.\\s","\\(","especifique","\\)"), "NA"))%>%#dplyr::count(banio_conectado, sort = T)
  #group_by(material_paredes, otro_material_paredes) %>% dplyr::count()
  unite(banio_conexion, banio_conectado, otro_banio_conectado) %>% 
  mutate(banio_conexion= str_replace(banio_conexion,"_NA", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(banio_conexion= str_replace(banio_conexion,"NA_", "")) 


#unap6a %>% glimpse()

#unap6a %>% dplyr::select(1, dplyr::contains("banio")) %>% glimpse()

unap6a %>% dplyr::count(banio_conexion, sort = T)
  #mutate(material_pared= fct_lump(material_pared, 3)) %>% dplyr::count(material_pared, sort = T)
  #group_by(trabajo, trabajo_paga_ultima) %>% dplyr::count(trabajo_paga_ultima, sort = T)
```

#### malaria

- `recibio_**tratamiento**_malaria` will be used in the next step

```{r, fig.width=8, fig.height=4,fig.align="center"}
unap6a <- unap6a %>% #str()
  #dplyr::select(1, age, residence, dplyr::contains("malaria")) %>% #str()
  mutate(recibio_tratamiento_malaria= str_replace(recibio_tratamiento_malaria,"\\d+\\.\\s", "")) %>% 
  mutate(recibio_tratamiento_malaria= str_replace(recibio_tratamiento_malaria,str_c("\\s","\\(","especifique","\\)"), "")) %>%
  mutate(epi_especie_causo_malaria= str_replace(epi_especie_causo_malaria,"\\d+\\.\\s", "")) %>% 
  mutate(epi_alguien_tuvo_malaria= str_replace(epi_alguien_tuvo_malaria,"\\d+\\.\\s", "")) #%>% #dplyr::count(epi_alguien_tuvo_malaria, sort = T)
  #dplyr::count(cuantas_tenido_malaria, sort = T)
  #dplyr::count(epi_meses_ultima_malaria, sort = T)
  #dplyr::count(cantidad_tratamiento_malaria, sort = T)

#unap6a %>% glimpse()

unap6a %>% 
  dplyr::select(1, age, residence, dplyr::contains("malaria")) %>%
  glimpse()

unap6a %>% 
  group_by(age, residence, cuantas_tenido_malaria) %>% #dplyr::count(sort = T)
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(age, cuantas_tenido_malaria, colour=residence)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="episodios de malaria")
```

#### tratamiento

- **`ndias_drogas_tratamiento_03`** is a inconsistent variable
- remove variables containing **`_tratamiento_03`** due to full of NA's and the shown inconsistence.


```{r}
unap6a %>% 
  group_by(drogas_tratamiento_03, ndias_drogas_tratamiento_03) %>% dplyr::count(sort = T)

unap6a <- unap6a %>% 
  #dplyr::select(-dplyr::contains("tratamiento_03")) %>% 
  #dplyr::select(1, dplyr::contains("tratamiento")) %>% #str() #dplyr::count(busco_ayuda_tratamiento)
  mutate(no_busco_ayuda_tratamiento= str_replace(no_busco_ayuda_tratamiento,
                                                 "\\d+\\.\\s", "")) %>% 
  #dplyr::count(no_busco_ayuda_tratamiento)
  mutate(donde_busco_tratamiento= str_replace(donde_busco_tratamiento,
                                              "\\d+\\.\\s", "")) %>% 
  #dplyr::count(donde_busco_tratamiento)
  mutate(respuesta_drogas_tratamiento=str_replace(respuesta_drogas_tratamiento,
                                                  "\\d+\\.\\s","")) # %>% 
  #dplyr::count(respuesta_drogas_tratamiento)
  #dplyr::count(drogas_tratamiento_01)
  #str()

unap6a %>% 
  dplyr::select(1, dplyr::contains("tratamiento")) %>% 
  glimpse()

#unap6a %>% dplyr::count(drogas_tratamiento_01, sort = T)
unap6a %>% 
  group_by(drogas_tratamiento_01, ndias_drogas_tratamiento_01) %>% 
  dplyr::count(sort = T)

#unap6a %>% dplyr::count(drogas_tratamiento_02, sort = T)
unap6a %>% 
  group_by(drogas_tratamiento_02, ndias_drogas_tratamiento_02) %>% 
  dplyr::count(sort = T)
```


#### enf

- remove **3** variables due to NA's
- `diabetes` and `cardiaca` with factor inconsistencies
- `enf_..._si` and `enf_..._no` were merge into `enf_...` for each disease
- remove **5** more variables after merging
- show a combination of chronic diseases frequencies

- inconsistency between `enfermedad_cronica` and `enf_...` variables

```{r}
unap6a <- unap6a %>% #str()
  dplyr::select(-enf_especifique_neurologica,
         -enf_especifique_pulmonar,
         -enf_especifique_cardiaca) %>% 
  mutate_at(vars(enf_diabetes_si:enf_renal_no), factor) %>% 
  #dplyr::select(1, dplyr::contains("enf")) %>% glimpse()
  
  mutate(enf_diabetes_si=fct_recode(enf_diabetes_si,"si"="1","no"="0"), 
         enf_diabetes_no=fct_recode(enf_diabetes_no,"no"="1","si"="0")) %>% 
  #dplyr::select(enf_diabetes_si,enf_diabetes_no) %>% count()
  unite(enf_diabetes, enf_diabetes_si, enf_diabetes_no, sep = "") %>% 
  #dplyr::count(enf_diabetes)
  mutate(enf_diabetes= str_replace_all(enf_diabetes, "nono","no")) %>% 
  #dplyr::count(enf_diabetes)
  mutate(enf_diabetes= str_replace_all(enf_diabetes, "..si|si..","si")) %>% 
  #dplyr::count(enf_diabetes)

  mutate(enf_neurologica_si=fct_recode(enf_neurologica_si,"si"="1","no"="0"),
         enf_neurologica_no=fct_recode(enf_neurologica_no,"no"="1","si"="0")) %>% 
  #dplyr::select(enf_neurologica_si,enf_neurologica_no) %>% count()
  unite(enf_neurologica, enf_neurologica_si, enf_neurologica_no, sep = "") %>% 
  #dplyr::count(enf_neurologica)
  mutate(enf_neurologica= str_replace_all(enf_neurologica, "nono","no")) %>% 
  #dplyr::count(enf_neurologica)
  mutate(enf_neurologica= str_replace_all(enf_neurologica, "..si|si..","si")) %>% 
  #dplyr::count(enf_neurologica)
    
  mutate(enf_pulmonar_si=fct_recode(enf_pulmonar_si,"si"="1","no"="0"), 
         enf_pulmonar_no=fct_recode(enf_pulmonar_no,"no"="1","si"="0")) %>% 
  #dplyr::select(enf_pulmonar_si,enf_pulmonar_no) %>% count()
  unite(enf_pulmonar, enf_pulmonar_si, enf_pulmonar_no, sep = "") %>% 
  #dplyr::count(enf_pulmonar)
  mutate(enf_pulmonar= str_replace_all(enf_pulmonar, "nono","no")) %>% 
  #dplyr::count(enf_pulmonar)
  mutate(enf_pulmonar= str_replace_all(enf_pulmonar, "..si|si..","si")) %>% 
  #dplyr::count(enf_pulmonar)
  
  mutate(enf_cardiaca_si=fct_recode(enf_cardiaca_si,"si"="1","no"="0"), 
         enf_cardiaca_no=fct_recode(enf_cardiaca_no,"no"="1","si"="0")) %>% 
  #dplyr::select(enf_cardiaca_si,enf_cardiaca_no) %>% count()
  unite(enf_cardiaca, enf_cardiaca_si, enf_cardiaca_no, sep = "") %>% 
  #dplyr::count(enf_cardiaca)
  mutate(enf_cardiaca= str_replace_all(enf_cardiaca, "nono|NAno|noNA","no")) %>% 
  #dplyr::count(enf_cardiaca)
  mutate(enf_cardiaca= str_replace_all(enf_cardiaca, "NANA",replacement = NA_character_)) %>% 
  #dplyr::count(enf_cardiaca)
  mutate(enf_cardiaca= str_replace_all(enf_cardiaca, "..si|si..","si")) %>% 
  #dplyr::count(enf_cardiaca)

  mutate(enf_renal_si=fct_recode(enf_renal_si,"si"="1","no"="0"), 
         enf_renal_no=fct_recode(enf_renal_no,"no"="1","si"="0")) %>% 
  #dplyr::select(enf_renal_si,enf_renal_no) %>% count()
  unite(enf_renal, enf_renal_si, enf_renal_no, sep = "") %>% 
  #dplyr::count(enf_renal)
  mutate(enf_renal= str_replace_all(enf_renal, "nono","no")) %>% 
  #dplyr::count(enf_renal)
  mutate(enf_renal= str_replace_all(enf_renal, "..si|si..","si")) 

unap6a %>% #dplyr::count(enf_renal)
  
  #dplyr::count(enf_especifique_renal)
  #dplyr::count(enf_especifique_otro)
  #dplyr::count(enf_tomo_medicinas_parasitos)
  #dplyr::count(enf_medicinas_parasitos)
  
  dplyr::select(1, dplyr::contains("enf")) %>% glimpse()

unap6a %>%
  dplyr::select(enfermedad_cronica, enf_diabetes, enf_neurologica, enf_pulmonar, enf_cardiaca, enf_renal) %>% 
  group_by(enfermedad_cronica, enf_diabetes, enf_neurologica, enf_pulmonar, enf_cardiaca, enf_renal) %>% count()
```

#### epi_

```{r}
unap6a <- unap6a %>%
  dplyr::select(-epi_especifique_viajo_fuera_villa) %>% ##dato crudo nulo (pocos reportan pucallpa)
  mutate(epi_tipo_fuente_agua= str_replace(epi_tipo_fuente_agua,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_tipo_fuente_agua)
  mutate(epi_distancia_fuente_agua= str_replace(epi_distancia_fuente_agua,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_distancia_fuente_agua)
  
  #mutate(epi_uso_repelente_mosquito= str_replace(epi_uso_repelente_mosquito,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_uso_repelente_mosquito)
  #mutate(epi_uso_mangas_largas= str_replace(epi_uso_mangas_largas,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_uso_mangas_largas)
  #mutate(epi_uso_ropa_impregnada= str_replace(epi_uso_ropa_impregnada,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_uso_ropa_impregnada)
  #mutate(epi_cercania_fuente_agua= str_replace(epi_cercania_fuente_agua,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_cercania_fuente_agua)
  #mutate(epi_estuvo_campo_antes= str_replace(epi_estuvo_campo_antes,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_estuvo_campo_antes)
  
  #mutate(epi_uso_redes_cama= str_replace(epi_uso_redes_cama,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_uso_redes_cama)
  #mutate(epi_duerme_ventanas_abiertas= str_replace(epi_duerme_ventanas_abiertas,"\\d+\\.\\s", "")) %>%
  ##dplyr::count(epi_duerme_ventanas_abiertas)
  #mutate(epi_duerme_cerca_monte= str_replace(epi_duerme_cerca_monte,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_duerme_cerca_monte)
  #mutate(epi_rocia_con_insecticida= str_replace(epi_rocia_con_insecticida,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_rocia_con_insecticida)
  mutate(epi_ultimas_semanas_viajo= str_replace(epi_ultimas_semanas_viajo,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_ultimas_semanas_viajo)
  
  mutate(epi_frecuencia_rocia_casa= str_replace(epi_frecuencia_rocia_casa,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_frecuencia_rocia_casa)
  
  #mutate(epi_estado_campos_agricultura= str_replace(epi_estado_campos_agricultura,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_estado_campos_agricultura)
  #mutate(epi_estado_canal_agua= str_replace(epi_estado_canal_agua,"\\d+\\.\\s", "")) %>% 
  ##dplyr::count(epi_estado_canal_agua)
  mutate(epi_adquirio_redes_dormir= str_replace(epi_adquirio_redes_dormir,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_adquirio_redes_dormir)
  
  mutate(epi_uso_red_dormir= str_replace(epi_uso_red_dormir,"\\d+\\.\\s", "")) %>% 
  #dplyr::count(epi_uso_red_dormir)
  mutate(epi_redes_usadas_actualmente= str_replace(epi_redes_usadas_actualmente,"\\d+\\.\\s", "")) 

unap6a %>%#dplyr::count(epi_redes_usadas_actualmente)
  
  #dplyr::count(epi_meses_ultima_malaria)
  #dplyr::count(epi_especie_causo_malaria)
  #dplyr::count(epi_cantidad_redes_usadas_actualmente)
  dplyr::select(1, dplyr::contains("epi_")) %>% glimpse()
```

#### nivel educacion

```{r}
unap6a <- unap6a %>% #dplyr::count(material_pared,material_piso)
  #dplyr::select(nivel_educacion) %>% #dplyr::count(nivel_educacion)
  mutate(nivel_educacion=str_replace(nivel_educacion,"Secundaria Completa|Tecnica(\\s)(.+)|Universidad(\\s)(.+)","Secundaria Completa / +")) %>%
  mutate(nivel_educacion=str_replace(nivel_educacion,"Primaria(\\s)(.+)","Primaria")) #%>%dplyr::count(nivel_educacion)
```

#### no patterns

```{r}
unap6a <- unap6a %>%
  dplyr::select(-email) %>% 
  dplyr::select(-fecha_actualizacion) %>%  
  mutate(combustible_cocinar= str_replace(combustible_cocinar,"\\d+\\.\\s", "")) %>% 
  mutate(pertenencia_casa= str_replace(pertenencia_casa,"\\d+\\.\\s", "")) %>% 
  mutate(tipo_casa= str_replace(tipo_casa,"\\d+\\.\\s", "")) %>% 
  mutate(fuente_agua= str_replace(fuente_agua,"\\d+\\.\\s", "")) %>% 
  mutate(relacion_jefe_familia= str_replace(relacion_jefe_familia,"\\d+\\.\\s", "")) 

unap6a %>%
  dplyr::select(1, 
         -dplyr::contains("ah_"),
         -dplyr::contains("otro_"),
         -dplyr::contains("trabajo"),
         -dplyr::contains("material"),
         -dplyr::contains("banio"),
         -dplyr::contains("malaria"),
         -dplyr::contains("tratamiento"),
         -dplyr::contains("enf"),
         -dplyr::contains("epi_")) %>% 
  dplyr::select(-starts_with("ah_")) %>%
  #dplyr::select(-email) %>% 
  #dplyr::select(-fecha_actualizacion) %>%
  #dplyr::count(estado_civil, sort = T)
  #dplyr::count(nivel_educacion, sort = T)
  #dplyr::count(estado, sort = T)
  glimpse()
```

## Merge datasets

- new DB names:
    + **`epid`**: current `unap6a` of Epidemiological variables **(963x88)**
    + **`unap`**: previous `unap0` **(960x16)** stored in `data/03-covariat-uap.rds` with:
        + undated `age`, `sex`, `community`
        + consensus `micro`, `pcr`
    + **`sero`**: previous `std.uty` **(969x5)** stored in `data/02-dtfilter-uty.rds` with: 
        + `Ab.units`, `mean.OD` per `Specie`

### rows differ

- all the datasets share **959 core ID's**.

TO DO:

- CHECK IF **intersections** outside the core **959 samples** have ID typos.

```{r, fig.align='center', fig.width=5, fig.height=5}
unapid <- unap0 %>% dplyr::select(id)      # 960
seroid <- std.uty %>% dplyr::select(id=ID) # 969
epidid <- unap6a %>% dplyr::select(id)     # 963

id_unap= unlist(as.list(unapid))
id_sero= unlist(as.list(seroid))
id_epid= unlist(as.list(epidid))

tmp <- gplots::venn(list(unap=id_unap,epid=id_epid,sero=id_sero))
```
**INTERSECTIONS**

- **epid** only
```{r}
attr(tmp, "intersections")$epid
```
- **sero** only
```{r}
attr(tmp, "intersections")$sero
```
- **unap:sero** only
```{r}
attr(tmp, "intersections")$`unap:sero`
```

### merge

- Merge **`unap + sero`** stored in `data/03-covariat-uty.csv` (`std.uty.cov`) 
with current **`epid`** (`unap6a`)
- This will be the final unified dataset.

dicoveries:

- **`epid`** and **`unap + sero`** differs in `age`, `sex`, `community`
    + `age`, `sex`, `community` were removed from **`epid`** 
    + logical vectors such as `_0` and `_n1` were removed from **`unap+sero`** 
    previous merging

```{r}
std.uty.cov.new <- std.uty.cov %>% 
  dplyr::select(-sex_0, 
                -micro_n1, 
                #-sex_1, 
                #-micro_1,
                #-pcr_1, 
                #-age_1, 
                -pcr_n1,
                -age_0) %>% 
  mutate(
    micro_1=fct_recode(micro_1,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3"),
    micro_n=fct_recode(micro_n,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3"),
    micro_c=fct_recode(micro_c,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3"),
    pcr_1=fct_recode(pcr_1,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3"),
    pcr_n=fct_recode(pcr_n,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3"),
    pcr_c=fct_recode(pcr_c,
                       "NG"="0","PV"="1","PF"="2","PV/PF"="3")
  )

unap6b <- unap6a %>% 
  dplyr::select(-age, -sex, -community) %>% 
  full_join(
    std.uty.cov.new
            , by="id") # 973
unap6b %>% glimpse()

unap6c <- unap6a %>% 
  dplyr::select(-age, -sex, -community) %>% 
  inner_join(
    std.uty.cov.new
             , by="id")# 959
#unap6c %>% glimpse()
```

## Data dictionary
```{r, message=FALSE, eval=TRUE}
datadic <- readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 2)#SPANISH
#datadic <- readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 3)#ENGLISH
whatyouwant <- setNames(as.character(datadic$`EXPLICACIÓN DE VARIABLE`), datadic$VARIABLE)
#http://stackoverflow.com/questions/19265172/converting-two-columns-of-a-data-frame-to-a-named-vector
unap6x <- Hmisc::upData(unap6c, labels = whatyouwant, force.single = F)
```
```{r}
unap6x <- Hmisc::upData(unap6x, 
                        labels = c(Ab.unit_Pfal="standardize AU Pfal",
                                   Ab.unit_Pviv="standardize AU Pviv",
                                   mean.OD_Pfal="raw mean OD Pfal",
                                   mean.OD_Pviv="raw mean OD Pviv",
                                   sex_4="gender-UNAP_UPDATE",
                                   age_1="age",
                                   age_4="age-UNAP_UPDATE",
                                   micro_1="microscopy UNAP",
                                   micro_n="microscopy NAMRU",
                                   micro_c="microscopy CONCENSUS",
                                   pcr_1="pcr UNAP",
                                   pcr_n="pcr NAMRU",
                                   pcr_c="pcr CONCENSUS"), 
                        force.single = F)
```

```{r}
#Hmisc::contents(unap6x)
Hmisc::html(Hmisc::contents(unap6x), maxlevels=10, levelType='table')
```


## Output

- a **`full`** dataset **(973x100)** was written as `data/04-epidemio-full.csv`.
- a **`core`** dataset **(959x100)** was written as `data/04-epidemio-core.csv`.

### write .csv .rds
```{r}
readr::write_csv(unap6b, "data/04-epidemio-full.csv")
saveRDS(unap6b, "data/04-epidemio-full.rds")

readr::write_csv(unap6c, "data/04-epidemio-core.csv")
readr::write_rds(unap6c, "data/04-epidemio-core.rds")
```

### write dta

- without labels due to errors in writing labelled variables

**.dta restrictions:**

- STATA requires labelled integers (not numeric variables)
- do not allow dots in variable name
- string length of variable names

```{r}
haven::write_dta(unap6c %>% 
                   dplyr::rename(Ab_unit_Pfal=Ab.unit_Pfal,
                                 Ab_unit_Pviv=Ab.unit_Pviv,
                                 mean_OD_Pfal=mean.OD_Pfal,
                                 mean_OD_Pviv=mean.OD_Pviv,
                                 epi_n_redes_adquirio_dormir=
                                   epi_cantidad_redes_adquirio_dormir,
                                 epi_n_redes_usadas_actualmente=
                                   epi_cantidad_redes_usadas_actualmente)
                 , "data/04-epidemio-core.dta", version=13)

#haven::read_dta("data/04-epidemio-core.dta")
#haven::read_dta("data/04-epidemio-core.dta") %>% haven::as_factor() # WORKS PERFECT!
```


To check the data in R, use the following commands:


```r
# load data
data <- readRDS("path-to-file/04-epidemio-core.rds")

# attach specialized package
library(tidyverse) 
# if not installed, run the following:
# install.packages("tidyverse")

# repetitive strings are recognized as factors (as shown in APPENDIX: Plots and problems)
# in order to ease summary output, coerce them to factors:
data %>% mutate_if(is.character, factor) %>% glimpse()  # all variables and class
data %>% mutate_if(is.character, factor) %>% str()      # all variables and structure
data %>% mutate_if(is.character, factor) %>% summary()  # all variables and descriptive stats
```

***

## APPENDIX: age updates {#slope}

Using `slopegraph` [@slopegraph]

```{r, fig.height=8, fig.width=6, fig.align='center'}
x <- unap6a %>% 
  dplyr::select(1,dplyr::contains("age")) %>% 
  full_join(std.uty.cov %>% 
              dplyr::select(1,dplyr::contains("age"), -age_0), 
            by="id") %>% 
  mutate(age_0= age==age_1 & age_1==age_4) %>% 
  filter(age_0==FALSE) %>% 
  dplyr::select(-age_0)

y <- x$id
z <- x %>% dplyr::select(-id) %>% as.data.frame()
rownames(z) <- y

slopegraph::ggslopegraph(z, 
                         offset.x = 0.06,
                         xlim = c(.5,3.5),
                         xlabels = c("unap6", "unap1", "unap4")) + theme_minimal() +
  labs(title="Age updates across datasets")
```


## APPENDIX: Plots and problems

```{r, fig.width=10, fig.height=4, fig.align="center"}
#
by_age <- unap6a %>%
  filter(!is.na(age)) %>%
  group_by(age, estado_civil) %>%
  dplyr::count() %>%
  mutate(prop= n / sum(n)) # counts marital inside each age level
#
a <- ggplot(by_age, aes(age, prop)) +
  geom_line(aes(colour= fct_reorder2(estado_civil, 
                                     age, prop))) +
  labs(colour= "Marital status",
       title= "Proportion of marital status per age",
       subtitle= "Data from a Epidemiological census in Iquitos, Loreto - Peru",
       caption= "a caption using ggplot() + labs()")

by_edu <- unap6a %>%
  filter(!is.na(age)) %>%
  group_by(age, nivel_educacion) %>%
  dplyr::count() %>%
  mutate(prop= n / sum(n)) # counts marital inside each age level

b <- ggplot(by_edu, aes(age, prop)) +
  geom_line(aes(colour= fct_reorder2(nivel_educacion, 
                                     age, prop))) +
  labs(colour= "Education level",
       title= "Proportion of education level per age",
       subtitle= "Data from a Epidemiological census in Iquitos, Loreto - Peru",
       caption= "a caption using ggplot() + labs()")

Rmisc::multiplot(a, b, cols=2)

```


***

### problems 3

#### malaria incidence plot

- check [above](#malaria) for a plot relating `age`, `residence` and `malaria`

```{r, fig.height=3, fig.width=6, fig.align="center"}
unap6a %>%
  ggplot(aes(cuantas_tenido_malaria)) + geom_histogram()

unap6a %>%
  group_by(tenido_malaria, opciones_tenido_malaria) %>%
  dplyr::summarise(n=n())

unap6a %>%
  group_by(tenido_malaria, opciones_tenido_malaria, cuantas_tenido_malaria) %>%
  dplyr::summarise(n=n()) %>%
  filter(tenido_malaria!="Si" | opciones_tenido_malaria!="Si")
```

#### date and time class

```{r}
unap6at <- unap6 %>% 
  dplyr::select(fecha_registro) %>% 
  mutate(fecha_registro= str_replace_all(fecha_registro, "\\/","-")) %>% 
  rownames_to_column(var = "row")

unap6at_1 <- unap6at %>% 
  tidyr::separate(fecha_registro, c("date", "time", "mer"), sep = " ") %>% filter(mer=="p.m.")
  #tidyr::separate(date, c("day", "month", "year"), sep = "-") #%>% count("year")
  
dmy_hms(unap6at$fecha_registro[1])
```




***

### sheet 2
```{r}
readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 2)
```


### sheet 3
```{r}
readxl::read_excel("data/UNAP-6-EPI-08feb2017.xlsx", sheet = 3)
```

<!--## References-->
