==> 05-epidb_up.Rmd <==
# NEW UPDATE {#upnuu}

## Aim

* Principal: 
    - Contrastar últimas actualizaciones (UNAP-7 y UNAP-8) con base Epidemiologica Consenso (UNAP-6c).
    - Generar __tablas resumen__ de:
        - Baseline characteristics of study population
        - Control de calidad (UNAP vs N6 microscopy and PCR results)

* Secondary: 
    - Check variable consistency across the database.

## Results

- Entre las dos últimas bases entregadas (UNAP-7 y UNAP-8)
    + `edad` es igual.
    + `sexo` difiere en __4__ individuos,
    + se prioriza a `sexo_consentimiento` en __UNAP-8__ (origilamente, `BASE_BARRIDO_2015-I_LIMPIA.xlsx`).
        - Cuenta con __"DATO VERIFICADO CON BARRIDO 2016"__

- Entre la última actualización y base concenso (UNAP-6)
    + `edad_DNI.7` corrige __61__ valores,
        - Ver contraste con `tiempo de residencia` y variación con respecto a la edad de bases previas.
    + `sexo_consentimiento` corrige __3__ valores,

- Valores actualizados de lectura de OD de UNAP-7 __no son tomados en cuenta__.
    + Aquellos valores fueron extraídos, estandarizados y filtrados directamente de los Templates (_data cruda_).
    

## Results II

- Detalles de inconsistencias (residencia y eventos previos de malaria)

- Adición de la información de georeferencia con mapeo descriptivo

- __Tabla 01__ de Baseline characteristics

- __Tabla 02__ de Malaria prevalence (by microscopy and PCR)

## To Do

<!--
- __Confirmar__ variables seelccionadas previo a la ejecución de la estadística inferencial. 
[Ver aquí](#selected-variables)
    + Ver resumen descriptivo. [Ver aquí](#resumen-descriptivo)

- Solve **inconsistencies**?
    + __17__ individuos tienen tiempo de residencia mayor a la edad última edad actualizada.
-->

- Ejecutar análisis de __seropositividad__ para generar __tabla resumen__ de:
    + Malaria seroprevalence by study site
- Ejecutar análisis __univariado y multivariado__ para generar __tabla resumen__ de:
    + Risk factors for species-specific malaria exposure
- Ejecutar __modelos de seroconversión__ para generar __gráfico resumen__ de:
    + Age-seroprevalence curves in each site
    + Rosas-Aguirre et al. 2015 lo restringe a individuos menores de 30 años. ¿Aplica para Zungarococha?

## Dependencies

The required R packages for this analysis are:

+ `tidyverse` [@tidyverse]
+ `readxl` [@readxl]
+ `forcats` [@forcats]
+ `stringr` [@stringr]

```{r, results='hide', message=FALSE}
##essential
#library(XLConnect)    # load EXCEL workbooks
#library(drc)          # Dose-Response modeling
##accesory
library("DiagrammeR") # method Flowchart
library("knitr")      # To display nice tables
library("tidyverse")
#ggplot2, tibble, tidyr, readr, purr, dplyyr
##extra
library("Rmisc")      # multiplot ggplots
library("haven")      # import STATA df
library("readxl")     # import EXCEL df
library("forcats")    # factor vectors
library("lubridate")  # for dates
library("stringr")    # for strings
library("viridis")    # color visualization
```

## Input

```{r}
# load data
unap6c <- readr::read_rds("data/04-epidemio-core.rds")

# attach specialized package
#library(tidyverse) 
# if not installed, run the following:
# install.packages("tidyverse")

# repetitive strings are recognized as factors (as shown in APPENDIX: Plots and problems)
# in order to ease summary output, coerce them to factors:

#unap6c %>% mutate_if(is.character, factor) %>% glimpse()  # all variables and class
#unap6c %>% mutate_if(is.character, factor) %>% str()      # all variables and structure
#unap6c %>% mutate_if(is.character, factor) %>% summary()  # all variables and descriptive stats

#unap6c
```

### Information of updates

```bash
#less README.md
```

### Standard file names

```bash
cp data-raw/RESULTADOS_BARRIDO_2015-I_FINAL_SETIEMBRE.xlsx \
data/UNAP-7-DB-25ago2017.xlsx

cp data-raw/RESULTADOS_BARRIDO_2015-I_FINAL_SETIEMBRE_2.xlsx \
data/UNAP-7-DB-19set2017.xlsx

cp data-raw/BASE_BARRIDO_2015-I_LIMPIA.xlsx \
data/UNAP-8-EPI-25ago2017.xlsx

cp data-raw/BASE_BARRIDO_2015-I_LIMPIA_2.xlsx \
data/UNAP-8-EPI-19set2017.xlsx
```

```bash
ls -shrt data/UNAP-*
```

### Tidy UNAP-7: RESULTADOS

```{r}
readxl::excel_sheets("data/UNAP-7-DB-25ago2017.xlsx")
```

```{r}
unap7_pr <- readxl::read_excel("data/UNAP-7-DB-25ago2017.xlsx", sheet = 1,na = ".") %>% 
  dplyr::rename(id=`CÓDIGO`,
                Pf_MSP1=`Pf_MSP1-QTSR`,
                Pv_MSP1=`Pv200_MSP1`) %>% 
  dplyr::mutate(id= str_replace_all(id, "_","-")) 

unap7 <- unap7_pr %>% #mejor cambialo a "_"
  dplyr::select(id,sexo_corregido,edad_DNI,Pv_MSP1,Pf_MSP1,densidad=Densidad_1er)
```

### Tidy UNAP-8: BASE BARRIDO

#### xlsx -> csv -> git

```r
readxl::read_excel("data/UNAP-8-EPI-25ago2017.xlsx", sheet = 1) %>% 
  write_csv("unap8.csv")
#readxl::read_excel("data/UNAP-8-EPI-19set2017.xlsx", sheet = 1)
```

```bash
git log --oneline
```

```r
readxl::read_excel("data/UNAP-8-EPI-19set2017.xlsx", sheet = 1) %>% 
  write_csv("unap8.csv")
```

```bash
git status
git diff unap8.csv
```

#### pre update

```{r}
readxl::excel_sheets("data/UNAP-8-EPI-25ago2017.xlsx")

readxl::read_excel("data/UNAP-8-EPI-25ago2017.xlsx", sheet = 2)
```

```{r}
unap8_pr <- readxl::read_excel("data/UNAP-8-EPI-25ago2017.xlsx", sheet = 1) %>% 
  dplyr::rename(id=id_integrante,
                residence="años_residencia") %>% 
  dplyr::mutate(id= str_replace_all(id, "_","-"),
                residence=as.numeric(residence))

#unap8 <- unap8_pr %>% #mejor cambialo a "_"
#  #select(id, contains("edad"), -enfermedad_cronica)
#  #select(id, contains("sexo"))
#  select(id, edad_DNI, sexo_consentimiento) #%>% summary()
```

#### pos update

```{r}
readxl::excel_sheets("data/UNAP-8-EPI-19set2017.xlsx")

readxl::read_excel("data/UNAP-8-EPI-19set2017.xlsx", sheet = 2)
```

```{r}
unap8_ps <- readxl::read_excel("data/UNAP-8-EPI-19set2017.xlsx", sheet = 1) %>% 
  dplyr::rename(id=id_integrante,
                residence="años_residencia") %>% 
  dplyr::mutate(id= str_replace_all(id, "_","-"),
                residence=as.numeric(residence))

#unap8x
unap8 <- unap8_ps %>% #mejor cambialo a "_"
  #select(id, contains("edad"), -enfermedad_cronica)
  #select(id, contains("sexo"))
  dplyr::select(id, edad_DNI, residence, sexo_consentimiento) #%>% #summary()
```

#### edad and residence test

- Se conservó la edad (DNI).
- Edición en el __tiempo de residencia__ en __02__ individuos: __ZG024_1__ y __ZG159_1__

```{r}
unap8_pr %>% full_join(unap8_ps, by="id", suffix=c(".8pr",".8ps")) %>% 
  mutate(edad_test= edad_DNI.8pr==edad_DNI.8ps,
         residence_test= residence.8pr==residence.8ps) %>% dplyr::count(edad_test,residence_test)
```

#### covariate test

- Corrección de la covariable `material_pared` por individuos de la misma casa: __LL076__ y __NN221__

```{r}
unap6c %>% filter(stringr::str_detect(id, "LL076-")) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria,
         Ab.unit_Pfal,Ab.unit_Pviv,
         micro_c,pcr_c
         )

unap8_pr %>% filter(stringr::str_detect(id, "LL076-")) %>% 
  dplyr::rename(material_pared=material_paredes) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria#,
         #Ab.unit_Pfal,Ab.unit_Pviv,
         #micro_c,pcr_c
         )
unap8_ps %>% filter(stringr::str_detect(id, "LL076-")) %>% 
  dplyr::rename(material_pared=material_paredes) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria#,
         #Ab.unit_Pfal,Ab.unit_Pviv,
         #micro_c,pcr_c
         )
```

```{r}
unap6c %>% filter(stringr::str_detect(id, "NN221-")) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria,
         Ab.unit_Pfal,Ab.unit_Pviv,
         micro_c,pcr_c
         )

unap8_pr %>% filter(stringr::str_detect(id, "NN221-")) %>% 
  dplyr::rename(material_pared=material_paredes) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria#,
         #Ab.unit_Pfal,Ab.unit_Pviv,
         #micro_c,pcr_c
         )
unap8_ps %>% filter(stringr::str_detect(id, "NN221-")) %>% 
  dplyr::rename(material_pared=material_paredes) %>% 
  dplyr::select(id, #sex_8, age_7, community_1,
         #nivel_educacion, trabajo, residence,
         electricidad_red_publica,
         material_pared, material_piso,
         epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
         epi_uso_red_dormir, #epi_redes_usadas_actualmente,
         tenido_malaria, cuantas_tenido_malaria#,
         #Ab.unit_Pfal,Ab.unit_Pviv,
         #micro_c,pcr_c
         )
```


#### household segmentation

##### count

```{r,eval=FALSE}
#unap8_pr %>% inner_join(unap8_ps,by = "id")

## TO SOLVE!!

unap8_ps %>% 
  dplyr::mutate(household= stringr::str_replace(id, "(.....)-(.+)","\\1")) %>% 
  #select(id,household,everything()) %>% 
  #dplyr::count(household) %>% 
  #filter(household=="LL003") %>% 
  dplyr::select(id,household,electricidad_red_publica:otro_banio_conectado,
         epi_cerca_fuente_agua:epi_distancia_fuente_agua) %>% 
  group_by(household) %>% 
  dplyr::summarise_all(unique=n_distinct(electricidad_red_publica))
```

## Between updates

### edad

- en EDAD: no hay incopatibilidad

```{r}
unap7 %>% full_join(unap8, by="id", suffix=c(".7",".8")) %>% 
  mutate(edad_test= edad_DNI.7==edad_DNI.8) %>% dplyr::count(edad_test)
```

### sexo

- en SEXO: difiere en __4__ individuos
    - ¿Cuál es el correcto?
    - `sexo_corregido` en UNAP-7
    - __`sexo_consentimiento` en UNAP-8 [CORRECTA]__
- posible error de digitación

```{r}
unap7 %>% full_join(unap8, by="id", suffix=c(".7",".8")) %>% 
  group_by(sexo_corregido,sexo_consentimiento) %>% dplyr::count()
```


## Contrast with UNAP-6

#### edad

- `edad_DNI.7` corrige __61__ valores.

```{r}
unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id") %>% 
  #select(contains("age")) %>% summary()
  #select(contains("edad"),-enfermedad_cronica) %>% summary()
  dplyr::mutate(edad_test= edad_DNI.7==age_4) %>% dplyr::count(edad_test)
  #select(id,contains("age"),contains("edad"),-enfermedad_cronica,residence) %>% 
  ##filter(edad_test=="FALSE") %>% 
  #dplyr::mutate(age_res_pre= age_4>=residence,
  #              age_res_up= edad_DNI.7>=residence) %>% 
  #group_by(age_res_pre,age_res_up) %>% dplyr::count()
```

#### residence time

- Contraste con `tiempo de residencia`.
    + __17__ individuos tienen tiempo de residencia mayor a la edad última edad actualizada. __[26set2017: CORREGIDO]__

```{r}
unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  mutate(residence_test= residence.6==residence.8) %>% 
  select(id,residence_test,residence.6,residence.8) %>% #filter(residence_test=="FALSE")
  dplyr::count(residence_test)
```


```{r, fig.height=3, fig.width=8}
a <- unap6c %>% 
  group_by(age_4, residence) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(age_4, residence, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="unap6-CORE")

b <- unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  group_by(edad_DNI.7, residence.8) %>%
  dplyr::summarise(n=n()) %>%
  mutate(prop= n/sum(n)) %>%
  ggplot(aes(edad_DNI.7, residence.8, colour=n)) + 
  geom_point() + 
  scale_color_viridis() +
  labs(title="unap8-UP")

Rmisc::multiplot(a,b,cols = 2)
```

- variación de la `edad` con respecto a la base de datos previa.

```{r, fig.height=8, fig.width=6, fig.align='center'}
x <- unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  dplyr::rename(age_7=edad_DNI.7) %>% 
  dplyr::select(1,dplyr::contains("age")) %>% 
  #full_join(std.uty.cov %>% 
  #            dplyr::select(1,dplyr::contains("age"), -age_0), 
  #          by="id") %>% 
  mutate(age_0= age_1==age_4 & age_1==age_7) %>% 
  filter(age_0==FALSE) %>% 
  dplyr::select(-age_0)

y <- x$id
z <- x %>% dplyr::select(-id) %>% as.data.frame()
rownames(z) <- y

slopegraph::ggslopegraph(z, 
                         offset.x = 0.06,
                         xlim = c(.5,3.5),
                         xlabels = c("unap1", "unap4", "unap7")) + theme_minimal() +
  labs(title="Age updates across datasets")
```


#### sexo

- OJO: justificar el uso de `sexo_consentimiento` __[STATUS: JUSTIFICADO]__
    - Cuenta con __"DATO VERIFICADO CON BARRIDO 2016"__
    - Menor diferencia con los datos de `unap-4`

```{r}
unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  dplyr::mutate(sexo_corregido= forcats::fct_recode(sexo_corregido,
                                                    "Female"="1",
                                                    "Female"="F",
                                                    "Male"="2"),
                sexo_consentimiento= forcats::fct_recode(sexo_consentimiento,
                                                         "Female"="F",
                                                         "Male"="M")
                ) %>% 
  #select(contains("sex")) %>% summary()
  #group_by(sexo_corregido,sexo_consentimiento) %>% dplyr::count()
  #mutate(sexo_test= sex_4==sexo_corregido) %>% dplyr::count(sexo_test)
  mutate(sexo_test= sex_4==sexo_consentimiento) %>% dplyr::count(sexo_test)
```

### mean OD reads

- NO USAR valores corregidos.
- VALORES EN TABLA provienen de la data cruda.

```{r}
unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  dplyr::mutate(pf_test= mean.OD_Pfal==Pf_MSP1,
                pv_test= mean.OD_Pviv==Pv_MSP1) %>% dplyr::count(pf_test,pv_test)
```


## Consensus

- __Seleccionar__ variables similares a la __tabla 1__ de Rosas-Aguirre et al. 2015
- Ejecutar __estadística inferencial__ para las proporciones entre comunidades.
- Ver __lista de todas las variables__ en la base de datos final:

```{r}
unap6c %>% 
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  dplyr::mutate(sexo_corregido= forcats::fct_recode(sexo_corregido,
                                                    "Female"="1",
                                                    "Female"="F",
                                                    "Male"="2"),
                sexo_consentimiento= forcats::fct_recode(sexo_consentimiento,
                                                         "Female"="F",
                                                         "Male"="M")
                ) %>% 
  select(-age_1,-age_4,-edad_DNI.8,-sex_1,-sex_4,-sexo_corregido,-Pf_MSP1,-Pv_MSP1) %>% 
  dplyr::rename(age_7=edad_DNI.7,
                sex_8=sexo_consentimiento) %>% glimpse()
```

### Selected variables

- __Confirmar__ su elección:
    + Ver resumen descriptivo. [Ver aquí](#resumen-descriptivo)
        - `edad` y `cuantas_tenido_malaria` serán categorizadas por cuartiles.
        - ¿Hay variable relacionada al __`material del techo`__? solo hay material de piso y paredes
        - ¿Hay variable relacionada __`tiempo desde la última exposición`__? 
        e.g. malaria en los últimos 12 meses.

```{r}
unap9c <- unap6c %>% #glimpse()
  inner_join(unap7 %>% 
              full_join(unap8, by="id", suffix=c(".7",".8")),
            by="id", suffix=c(".6",".8")) %>% 
  dplyr::mutate(sexo_corregido= forcats::fct_recode(sexo_corregido,
                                                    "Female"="1",
                                                    "Female"="F",
                                                    "Male"="2"),
                sexo_consentimiento= forcats::fct_recode(sexo_consentimiento,
                                                         "Female"="F",
                                                         "Male"="M")
                ) %>% 
  dplyr::select(-age_1,-age_4,-edad_DNI.8,-sex_1,-sex_4,-sexo_corregido,-Pf_MSP1,-Pv_MSP1,-residence.6) %>% 
  dplyr::rename(age_7=edad_DNI.7,
                sex_8=sexo_consentimiento,
                residence=residence.8) #%>% #glimpse()
  #dplyr::select(id, sex_8, age_7, community_1,
  #       nivel_educacion, trabajo, residence,
  #       electricidad_red_publica,
  #       material_pared, material_piso,
  #       epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
  #       epi_uso_red_dormir, epi_redes_usadas_actualmente, #mosquitero impregnado -> BUT CHECK ISSUE
  #       actualmente_estudiando,
  #       tenido_malaria, cuantas_tenido_malaria, densidad,
  #       Ab.unit_Pfal,Ab.unit_Pviv,
  #       mean.OD_Pfal, mean.OD_Pviv,
  #       micro_c,pcr_c
  #       )

unap9c %>% glimpse()
table(unap6c$epi_uso_red_dormir,unap6c$epi_redes_usadas_actualmente) #issue!!!!
```

##### Resumen descriptivo

```{r}
unap9c %>% mutate_if(is.character, factor) %>% summary()
```

#### Exploratorio

```{r}
unap9c
```

## Georeference data

- Unir y actualizar `unap9c` con los datos de georeferencia.
- __pendiente:__ visualizarlo mediante google maps.

```{r}
unapgr <- readxl::read_xlsx("data-raw/UNAP_georeference.xlsx",n_max=940) %>% 
              dplyr::rename(id=`CÓDIGO`) %>% 
              rename_all(funs(stringr::str_to_lower(.))) %>% 
              dplyr::mutate(id= str_replace_all(id, "_","-"))

unap9c <- unap9c %>% 
  inner_join(unapgr,
            by = "id")
```

## UPDATE: with unap8 covariates

```{r}
#unap8_ps %>% glimpse()

unap8_ps.ty <- unap8_ps %>% 
  mutate(tipo_trabajo= str_replace(tipo_trabajo,"\\d+\\.\\s", "")) %>% #count("tipo_trabajo")
  mutate(tipo_trabajo= str_replace(tipo_trabajo,"Actualmente ", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(tipo_trabajo= str_replace(tipo_trabajo,str_c("Otro ","\\(","especificar","\\)"), "NA")) %>% #dplyr::count(tipo_trabajo, sort = T)
  unite(trabajo, tipo_trabajo, otro_tipo_trabajo) %>% 
  mutate(trabajo= str_replace(trabajo,"_NA", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(trabajo= str_replace(trabajo,"NA_", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(trabajo= str_replace(trabajo,"_", " ")) %>% #dplyr::count(trabajo)
  mutate(trabajo= str_replace(trabajo,"NA", NA_character_)) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "(.+)(\\s)casa(\\s)(.+)|(.+)(\\s)casa","Ama de casa"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "(.+)(\\()estudiante(\\))(.+)|estudiante","Estudiante"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "Agricultor|agricultor","Agricultor"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "Granjero|granjero","Granjero"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "(.+)(\\()vigilante(\\))(.+)|vigilante|Vigilante","Vigilante"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "(.+)(\\()obrero(\\))|obrero(\\s)(.+)|obrero|obreri|albanil|albañil|construccion|contruccion","Obrero/albanil/construccion"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "Comerciante|comerciante","Comerciante"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "Chofer|chofer","Chofer"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "administrativo(\\s)(.+)|limpieza|personal(\\s)(.+)|Profesor","Administrativo/limpieza/profesor"
                                 )) %>% 
  mutate(trabajo=str_replace(trabajo,
                                 "mecanico|Mecanico|tecnico(\\s)(.+)","Mecanico/tecnico"
                                 )) %>% 
  #dplyr::count(trabajo) %>% print(n=Inf)


  mutate(trabajo_tdy=trabajo) %>% 
  mutate(trabajo_tdy=str_replace(trabajo_tdy,
                                 "(.+)(\\s)casa(\\s)(.+)|(.+)(\\s)casa","Housewife"
                                 )) %>% 
  mutate(trabajo_tdy=str_replace(trabajo_tdy,
                                 "(.+)(\\()estudiante(\\))(.+)|estudiante|pensionista","None (Students/Pensioner)"
                                 )) %>% 
  mutate(trabajo_tdy=str_replace(trabajo_tdy,
                                 "Agricultor|agricultor|Ganadero|Granjero|granjero|(.+)(\\()vigilante(\\))(.+)|vigilante|Vigilante|carpintero|pescador","Farmer/Guard/Logger/Fisher"
                                 )) %>% 
  mutate(trabajo_tdy=str_replace(trabajo_tdy,
                                 "(.+)(\\()obrero(\\))|obrero(\\s)(.+)|obrero|obreri|albanil|albañil|construccion|contruccion|Comerciante|comerciante|panadero|mototaxista|Chofer|chofer|guia(\\s)(.+)|(.+)(\\s)salud|administrativo(\\s)(.+)|limpieza|mecanico|Mecanico|personal(\\s)(.+)|Profesor|tecnico(\\s)(.+)","Laborer/Trader/Driver"
                                 )) %>% 
  #dplyr::count(trabajo_tdy,sort = T)
  mutate(trabajo_rpl=if_else(trabajo_tdy=="Farmer/Guard/Logger/Fisher",
                             "Outdoor",
                             "Other")) %>% 
  #mutate(trabajo_lmp= fct_lump(trabajo, 5)) %>% 
  #dplyr::count(trabajo_tdy) %>% arrange(desc(n))
  #dplyr::select(1, dplyr::contains("material")) %>% 
  mutate(material_piso= str_replace(material_piso,"\\d+\\.\\s", "")) %>% #dplyr::count(material_piso, sort = T)
  mutate(material_piso= str_replace(material_piso,"(.+)madera(.+)|(.+)(\\s)madera","madera")) %>% 
  mutate(material_paredes= str_replace(material_paredes,"\\d+\\.\\s", "")) %>% #dplyr::count(material_paredes, sort = T)
  mutate(material_paredes= str_replace(material_paredes,"\\so\\s", "_")) %>% #dplyr::count(material_paredes, sort = T)
  mutate(material_paredes= str_replace(material_paredes,str_c("Otro ","\\(","especifique","\\)"), "NA")) %>%#dplyr::count(material_paredes, sort = T)
  #group_by(material_paredes, otro_material_paredes) %>% dplyr::count()
  unite(material_pared, material_paredes, otro_material_paredes) %>% 
  mutate(material_pared= str_replace(material_pared,"_NA", "")) %>% #dplyr::count(trabajo, sort = T)
  mutate(material_pared= str_replace(material_pared,"NA_", "")) %>% 
  mutate(material_pared= str_replace(material_pared,"Madera(.+)","Madera")) %>% 
  mutate(material_pared= str_replace(material_pared,"(.+)_cemento_(.+)|(.+)_cemento","cemento")) %>% 
  mutate(nivel_educacion=str_replace(nivel_educacion,"Secundaria Completa","Secundaria Completa")) %>%
  mutate(nivel_educacion=str_replace(nivel_educacion,"Tecnica(\\s)(.+)","Tecnica")) %>%
  mutate(nivel_educacion=str_replace(nivel_educacion,"Universidad(\\s)(.+)","Universitaria")) %>%
  mutate(nivel_educacion=str_replace(nivel_educacion,"Primaria(\\s)(.+)","Primaria")) %>% 
  mutate(epi_uso_redes_cama= str_replace(epi_uso_redes_cama,"\\d+\\.\\s", "")) %>% 
  mutate(epi_uso_red_dormir= str_replace(epi_uso_red_dormir,"\\d+\\.\\s", "")) #%>% ######### DIFERENCE HERE!!!
  #dplyr::select(id, #sex_8, age_7, community_1,
  #       nivel_educacion, trabajo,trabajo_tdy,trabajo_rpl, residence,
  #       electricidad_red_publica,
  #       material_pared, material_piso,
  #       epi_uso_redes_cama, #epi_adquirio_redes_dormir, 
  #       epi_uso_red_dormir, epi_redes_usadas_actualmente, #mosquitero impregnado -> BUT CHECK ISSUE
  #       actualmente_estudiando,
  #       tenido_malaria, cuantas_tenido_malaria#,
         #Ab.unit_Pfal,Ab.unit_Pviv,
         #micro_c,pcr_c
  #       ) #%>% dplyr::count(trabajo)
#unap9c %>% dim()
#unap9c %>% glimpse()

#unap9c %>% select(id:community_1,Ab.unit_Pfal:longitud)

unap9d <- unap8_ps.ty %>% #glimpse()#dplyr::select(contains("sex"))#glimpse()
  dplyr::select(id,#residence, sexo_consentimiento,
                trabajo,
                trabajo_tdy,trabajo_rpl,material_piso,#material_paredes,
                material_pared,nivel_educacion#,epi_uso_redes_cama,epi_uso_red_dormir
                ) %>% #dplyr::count(trabajo,trabajo_tdy,trabajo_rpl) %>% print(n=Inf) #glimpse()
  inner_join(unap9c, by="id", suffix=c(".8",".9")) %>% #dplyr::select(id,contains("epi_uso_red")) %>% print(n=Inf) #glimpse()
  dplyr::select(-ends_with(".9")) %>% #OJO, PRIORIZA  sobre 9, ¿pero en cuales?
  rename_all(
      funs(
        #stringr::str_to_lower(.) %>%
        stringr::str_replace_all(., "\\.8", '')
      )
  ) %>%
  mutate(micro_c=forcats::fct_recode(micro_c,
                                     non="NG",
                                     #non="NA",
                                     viv="PV",
                                     fal="PF"),
         pcr_c=forcats::fct_recode(pcr_c,
                                   non="NG",
                                   #non="NA",
                                   viv="PV",
                                   fal="PF",
                                   mix="PV/PF")
         ) %>% 
  
  replace_na(list(micro_c = "non", pcr_c = "non")) %>%  #DO NOT REPLACE NA!!(2018-05-14) --> TAKE NOTE OF REASON
  
  #mutate(nivel_educacion= stringr::str_to_lower(nivel_educacion),
   #      trabajo= stringr::str_to_lower(trabajo),
  #       residence= stringr::str_to_lower(residence),
  #       material_pared= stringr::str_to_lower(material_pared),
  #       material_piso= stringr::str_to_lower(material_piso),
  #       epi_uso_redes_cama= stringr::str_to_lower(epi_uso_redes_cama),
  #       epi_uso_red_dormir= stringr::str_to_lower(epi_uso_red_dormir),
  #       tenido_malaria= stringr::str_to_lower(tenido_malaria)
  #       ) %>% 
  mutate(material_pared=stringr::str_replace(material_pared,"(.+)_(.+)_(.+)","\\1, \\2 y \\3")) %>% 
  mutate(material_pared=stringr::str_replace(material_pared,"(.+)_(.+)","\\1 y \\2")) %>% 
  mutate(material_piso=stringr::str_replace(material_piso,"(.+)\\/(.+)","\\1 o \\2")) %>%
  mutate(material_pared=stringr::str_replace(material_pared,"tripleyy","tripley")) %>% 
  mutate(
    nivel_educacion=as.factor(nivel_educacion),
    electricidad_red_publica=as.factor(electricidad_red_publica),
    material_pared=as.factor(material_pared),
    material_piso=as.factor(material_piso),
    epi_uso_redes_cama=as.factor(epi_uso_redes_cama),
    epi_uso_red_dormir=as.factor(epi_uso_red_dormir),
    tenido_malaria=as.factor(tenido_malaria)
         ) %>% 
  mutate(community_1=as.character(community_1)) %>% 
  mutate_if(is.factor,
    funs(
      stringr::str_to_lower(.)
    )
  ) %>% 
  dplyr::select(id,sex_8,age_7,community_1,everything()) %>% 
  mutate_if(is.character,
            funs(
      as.factor(.)
    )) %>% 
  dplyr::mutate(id=as.character(id))

unap9d %>% glimpse()
#unap9c %>% summary()
unap9d %>% summary()
```


## Output

```{r, fig.align='center', fig.width=5, fig.height=5}
unap6i <- unap6c %>% dplyr::select(id)
unap7i <- unap7 %>% dplyr::select(id)
unap9i <- unap9d %>% dplyr::select(id)
unapgi <- unapgr %>% dplyr::select(id)

i9_unap= unlist(as.list(unap9i))
i7_sero= unlist(as.list(unap7i))
ig_epid= unlist(as.list(unapgi))
i6_unap= unlist(as.list(unap6i))

tmp <- gplots::venn(list(unap9d=i9_unap,unap_geo=ig_epid,unap7=i7_sero,unap6c=i6_unap))
```

- a **`core`** dataset **(936x100)** was written as `data/05-epidemio-core.csv`.
    + __23__ samples from `unap6c` are lost.
    + __01__ sample from `unap7` is lost due to absence in core DB `unap6c`.

### write .csv .rds
```{r}
#readr::write_csv(unap6b, "data/04-epidemio-full.csv")
#saveRDS(unap6b, "data/04-epidemio-full.rds")

#unap9c 
#unap9d %>% glimpse()
readr::write_csv(unap9d, "data/05-epidemio-core.csv")
readr::write_rds(unap9d, "data/05-epidemio-core.rds")
```

### write dta

- without labels due to errors in writing labelled variables

**.dta restrictions:**

- STATA requires labelled integers (not numeric variables)
- do not allow dots in variable name
- string length of variable names

```{r}
haven::write_dta(unap9d %>% 
                   dplyr::rename(Ab_unit_Pfal=Ab.unit_Pfal,
                                 Ab_unit_Pviv=Ab.unit_Pviv,
                                 mean_OD_Pfal=mean.OD_Pfal,
                                 mean_OD_Pviv=mean.OD_Pviv#,
                                 #epi_n_redes_adquirio_dormir=
                                  # epi_cantidad_redes_adquirio_dormir,
                                 #epi_n_redes_usadas_actualmente=
                                  # epi_cantidad_redes_usadas_actualmente
                                 )
                 , "data/05-epidemio-core_stata13.dta", version=13)

haven::write_dta(unap9d %>% 
                   dplyr::rename(Ab_unit_Pfal=Ab.unit_Pfal,
                                 Ab_unit_Pviv=Ab.unit_Pviv,
                                 mean_OD_Pfal=mean.OD_Pfal,
                                 mean_OD_Pviv=mean.OD_Pviv#,
                                 #epi_n_redes_adquirio_dormir=
                                  # epi_cantidad_redes_adquirio_dormir,
                                 #epi_n_redes_usadas_actualmente=
                                  # epi_cantidad_redes_usadas_actualmente
                                 )
                 , "data/05-epidemio-core_stata14.dta", version=14)

#haven::read_dta("data/04-epidemio-core.dta")
#haven::read_dta("data/04-epidemio-core.dta") %>% haven::as_factor() # WORKS PERFECT!
```


To check the data in R, use the following commands:


```r
# load data
data <- readRDS("path-to-file/04-epidemio-core.rds")

# attach specialized package
library(tidyverse) 
# if not installed, run the following:
# install.packages("tidyverse")

# repetitive strings are recognized as factors (as shown in APPENDIX: Plots and problems)
# in order to ease summary output, coerce them to factors:
data %>% mutate_if(is.character, factor) %>% glimpse()  # all variables and class
data %>% mutate_if(is.character, factor) %>% str()      # all variables and structure
data %>% mutate_if(is.character, factor) %>% summary()  # all variables and descriptive stats
```


## Inconsistencies

### malaria events

- __01__ individuo reporta 1 evento de malaria pero se registra sin `tenido_malaria`.
- ¿Corrijo priorizando `cuantas_tenido_malaria`?

```{r}
unap9d %>% 
  select(tenido_malaria,cuantas_tenido_malaria) %>% 
  group_by(tenido_malaria,cuantas_tenido_malaria) %>% dplyr::count()
```

### On residence higher than age

- __04__ son __vivax+__ por PCR
- __05__ poseen __Ab.units contra P.vivax__ mayor al tercer cuartil.

```{r}
unap9d %>% 
  select(id, age_7,residence,micro_c,pcr_c) %>% 
  filter(residence>age_7) %>% 
  group_by(micro_c,pcr_c) %>% dplyr::count()

unap9d %>% 
  select(id, age_7,residence,pcr_c,Ab.unit_Pviv) %>% 
  filter(residence>age_7,Ab.unit_Pviv>quantile(Ab.unit_Pviv,0.75,na.rm=T))

summary(unap9d$Ab.unit_Pviv)
```

- lista de __17__ individuos con `tiempo de residencia` mayor a la edad última edad actualizada: __[26set2017: CORREGIDO]__

```{r}
unap9d %>% 
  select(id, age_7,residence,micro_c,pcr_c) %>% 
  filter(residence>age_7)
```

<!--
### Microscopy and PCR reads

- la 1era base con estos datos fue `unap5 (UNAP-5-QC-08feb2017.xlsx)`
- `unap5` solo presenta __222__ muestras con datos para micros y pcr
- la última base entregada `RESULTADOS_BARRIDO_2015-I_FINAL_SETIEMBRE.xlsx (UNAP-7-DB-25ago2017.xlsx)` 
rellenó las casillas sin datos con valores `0`.
- `micro_c` y `pcr_c` son las variables consenso obtenidas al __comparar__ UNAP vs N6 y __priorizar__ las lecturas _positivas_ obtenidas en N6.
    + sufijo `_7` son datos de `unap7` (`_a` y `_b`, 1era y 2da lectura),
    + sufijo `_1` son datos de UNAP en `unap5`, la 1era base con esta data.
    + sufijo `_n` son datos de NAM6 en `unap5`.

```{r}
# load data
readxl::read_xlsx("data/UNAP-5-QC-08feb2017.xlsx") %>% 
  dplyr::rename(id=CODE) %>% dplyr::mutate(db="unap5") %>%  dplyr::count(db)

unap_qc76 <- readxl::read_excel("data/UNAP-7-DB-25ago2017.xlsx", 
                              sheet = 1,na = ".") %>% 
  dplyr::rename(id=`CÓDIGO`,
                Pf_MSP1=`Pf_MSP1-QTSR`,
                Pv_MSP1=`Pv200_MSP1`,
                micro_7a=`1er micros (nolberto_anibal)`,
                micro_7b=`2da micros  (nolberto)`,
                pcr_pv=`PCR_Pv (Rubio et al 1999)`,
                pcr_pf=`PCR_Pf (Rubio et al 1999)`) %>% 
  dplyr::mutate(id= str_replace_all(id, "_","-")) %>% 
  select(id,micro_7a,micro_7b,pcr_pv,pcr_pf) %>% 
  inner_join(unap6c %>% 
              select(id,micro_c:pcr_n),
            by="id")

unap_qc76 %>% 
  select(contains("micro")) %>% 
  dplyr::mutate(micro_c=as.character(micro_c)) %>% 
  replace_na(list(micro_c="na"))  %>% 
  filter(micro_c=="na") %>% 
  group_by(micro_7a,micro_7b,micro_c,micro_1,micro_n) %>%  dplyr::count()

unap_qc76 %>% 
  select(contains("pcr")) %>% 
  dplyr::mutate(pcr_7=pcr_pv+pcr_pf) %>% 
  dplyr::mutate(pcr_c=as.character(pcr_c)) %>% 
  replace_na(list(pcr_c="na"))  %>% 
  filter(pcr_c=="na") %>% 
  group_by(pcr_7,pcr_c,pcr_1,pcr_n) %>%  dplyr::count()
```

- los valores negativos `NG` también fueron reemplazados por `0` en la última actualización:

```{r}
unap_qc76 %>% 
  select(contains("micro")) %>% 
  dplyr::mutate(micro_c=as.character(micro_c)) %>% 
  replace_na(list(micro_c="na"))  %>% 
  filter(micro_c=="NG") %>% 
  group_by(micro_7a,micro_7b,micro_c,micro_1,micro_n) %>%  dplyr::count()

unap_qc76  %>% 
  select(contains("pcr")) %>% 
  dplyr::mutate(pcr_7=pcr_pv+pcr_pf) %>% 
  dplyr::mutate(pcr_c=as.character(pcr_c)) %>% 
  replace_na(list(pcr_c="na"))  %>% 
  filter(pcr_c=="NG") %>% 
  group_by(pcr_7,pcr_c,pcr_1,pcr_n) %>%  dplyr::count()
```

- Todos los valores positivos
- Los resultados del __control de calidad (QC)__ fueron reportados anteriormente [aquí](#covar)

```{r}
unap_qc76 %>% 
  select(contains("micro")) %>% 
  dplyr::mutate(micro_c=as.character(micro_c)) %>% 
  replace_na(list(micro_c="na"))  %>% 
  filter(micro_c!="NG" & micro_c!="na") %>% 
  group_by(micro_7a,micro_7b,micro_c,micro_1,micro_n) %>%  dplyr::count()

unap_qc76  %>% 
  select(contains("pcr")) %>% 
  dplyr::mutate(pcr_7=pcr_pv+pcr_pf) %>% 
  dplyr::mutate(pcr_c=as.character(pcr_c)) %>% 
  replace_na(list(pcr_c="na"))  %>% 
  filter(pcr_c!="NG" & pcr_c!="na") %>% 
  group_by(pcr_7,pcr_c,pcr_1,pcr_n) %>%  dplyr::count()
```
-->
## Result tables

### Baseline characteristics of the study population

```{r}
unap9d_bc <- unap9d %>% 
  select(-residence,-(Ab.unit_Pfal:longitud)) %>% 
  dplyr::mutate(age_7=cut(age_7,c(0,
                                  quantile(unap9d$age_7,0.25,na.rm = T),
                                  quantile(unap9d$age_7,0.50,na.rm = T),
                                  quantile(unap9d$age_7,0.75,na.rm = T),
                                  Inf)),
                cuantas_tenido_malaria=cut(cuantas_tenido_malaria,
                                           c(-Inf,
                                             0,
                                             1,
                                             quantile(unap9d$cuantas_tenido_malaria,0.75,na.rm = T),
                                             Inf))
                )
```

```{r}
s1 <- Hmisc::summaryM(sex_8 + age_7 +
                        #community_1 +
                        nivel_educacion +
                        trabajo + trabajo_tdy + trabajo_rpl +
                        electricidad_red_publica +
                        material_pared +
                        material_piso +
                        #epi_uso_redes_cama +
                        epi_uso_red_dormir + tenido_malaria +
                        cuantas_tenido_malaria
                      ~ community_1 #+ 
                      #episodio_previo
                      ,
               data=unap9d_bc,
               overall=FALSE, test=TRUE)

Hmisc::latex(s1, caption='Baseline characteristics of the study population',
      exclude1=TRUE, npct='both', 
      #digits=2, eps = 0.01,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = TRUE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```

### Malaria prevalence by study site

- prevalencia en base a microscopía y PCR.
- __pendiente:__ seroprevalencia
- ¿Se tomará en cuenta _Incidence rate_? (ver Rosas-Aguirre et al. 2015)

```{r}
unap9d_pr <- unap9d %>% 
  select(id,community_1,Ab.unit_Pfal:pcr_c) %>% 
  #filter(micro_c!="NG" | pcr_c!="NG")
  dplyr::mutate(micro_c=stringr::str_replace(micro_c,"NG", 
                                             replacement = NA_character_),
                pcr_c=stringr::str_replace(pcr_c,"NG", 
                                             replacement = NA_character_))
```

```{r}
s1 <- Hmisc::summaryM(micro_c + pcr_c #+ Ab.unit_Pviv + Ab.unit_Pfal
                      ~ community_1 #+ 
                      #episodio_previo
                      ,
               data=unap9d_pr,
               overall=FALSE, test=TRUE)

Hmisc::latex(s1, caption='Malaria prevalence by study site',
      exclude1=FALSE, npct='both', 
      digits=2, #eps = 0.01,
      #prmsd=TRUE, brmsd=TRUE, msdsize=mu$smaller2, #NOT-EVALUATE if PDF
      middle.bold=TRUE, long = TRUE,
      #legend.bottom = TRUE, #insert.bottom = TRUE, 
      what="%", html = TRUE
      ) #change here for LaTeX PDF
```
